### 2022-04-04 #170 @whedon
[PRE REVIEW]: The Argo Online School: An e-learning tool to get started with Argo
**Submitting author:** @Alberto-GS (<a href="http://orcid.org/0000-0001-5781-9330">Alberto González</a>)
**Repository:** <a href="https://github.com/euroargodev/argoonlineschool" target ="_blank">https://github.com/euroargodev/argoonlineschool</a>
**Version:** 1.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
