+++ README.md
@@ -1,7 +1,3 @@
-[![Jupyter Book Badge](https://jupyterbook.org/badge.svg)](<https://euroargodev.github.io/argoonlineschool>)
-
-<img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/logoAoS.png" alt="argopy logo" width="100"/>
-
-# Argo Online School
-
-The Argo Online school is a set of videos and hands-on jupyter notebooks to describe the minimum requiremets to use, and understand, the Argo data. This is the source repository for https://euroargodev.github.io/argoonlineschool.
+|<img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/logoAoS.png" alt="argopy logo" width="100"/><br> The ``Argo Online School``  is a set of videos and hands-on jupyter notebooks to describe the minimum requiremets to use, and understand, the Argo data. <br> This is the source repository for https://euroargodev.github.io/argoonlineschool|
+|:---------:|
+|[![status](https://jose.theoj.org/papers/b66eaed8751b3adb6f2f4ad146380818/status.svg)](https://jose.theoj.org/papers/b66eaed8751b3adb6f2f4ad146380818) [![Jupyter Book Badge](https://jupyterbook.org/badge.svg)](<https://euroargodev.github.io/argoonlineschool>)|
