+++ _sources/intro.md
@@ -2,6 +2,7 @@
     <img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/logoAoS_banner.png" alt="AoS logo" width="600"/>
 </p>
 
+[![status](https://jose.theoj.org/papers/b66eaed8751b3adb6f2f4ad146380818/status.svg)](https://jose.theoj.org/papers/b66eaed8751b3adb6f2f4ad146380818) [![Jupyter Book Badge](https://jupyterbook.org/badge.svg)](<https://euroargodev.github.io/argoonlineschool>)
 
 # The Argo Online School
 Argo is an international program that collects information from inside the ocean using a fleet of floats that drift with the ocean currents. The floats move up and down between the surface and a mid-water level measuring ocean variables but spend almost all their life below the surface. Argo has been collecting more than 100,000 profiles per year since 2012, and nowadays [Argo](https://argo.ucsd.edu/) is the major component of the Global Ocean Observing System.
