### 2022-04-06 73a50e0
Create license.md
