+++ license.md
@@ -1,0 +1,171 @@
+## License for this book
+All the content in the Argo Online School (AOS) is licensed under the [Creative Commons Attribution-ShareAlike 4.0 International](https://creativecommons.org/licenses/by-sa/4.0/) (CC BY-SA 4.0) license.
+
+Terms of Use
+Creative Commons Master Terms of Use
+
+Effective as of 26 August 2020
+1. General Information Regarding These Terms of Use
+
+Master terms: Welcome, and thank you for your interest in Creative Commons (“Creative Commons,” “CC,” “we,” “our,” or “us”). Unless otherwise noted on a particular site or service, these master terms of use (“Master Terms”) apply to your use of all of the websites that Creative Commons Corporation operates. These include https://creativecommons.org, https://wiki.creativecommons.org, https://network.creativecommons.org, https://search.creativecommons.org, https://labs.creativecommons.org, http://openpolicynetwork.org, http://open4us.org, https://rightsback.org, http://teamopen.cc, and http://thepowerofopen.org, together with all other subdomains thereof, (collectively, the “Websites”). The Master Terms also apply to all products, information, and services provided through the Websites, including without limitation the license chooser, legal tools, the CCID service available at https://login.creativecommons.org/login (“CC Login Service”), and the CC Global Network community website (together with the Websites, the “Services”).
+
+Creative Commons operates opencovidpledge.org, but that website is governed by its own terms of use and privacy policy. Creative Commons does not operate the website at http://ccmixter.org.
+
+Additional terms: In addition to the Master Terms, your use of any Services may also be subject to specific terms applicable to a particular Service (“Additional Terms”). If there is any conflict between the Additional Terms and the Master Terms, then the Additional Terms apply in relation to the relevant Service.
+
+Collectively, the Terms: The Master Terms, together with any Additional Terms, form a binding legal agreement between you and Creative Commons in relation to your use of the Services. Collectively, this legal agreement is referred to below as the “Terms.”
+
+Human-readable summary of Sec 1: These terms, together with any special terms for particular websites, create a contract between you and Creative Commons. The contract governs your use of all websites operated by Creative Commons, unless a particular website indicates otherwise. These human-readable summaries of each section are not part of the contract, but are intended to help you understand its terms.
+2. Your Agreement to the Terms
+
+BY CLICKING “I ACCEPT” OR OTHERWISE ACCESSING OR USING ANY OF THE SERVICES (INCLUDING THE LICENSES, PUBLIC DOMAIN TOOLS, AND CHOOSERS), YOU ACKNOWLEDGE THAT YOU HAVE READ, UNDERSTOOD, AND AGREED TO BE BOUND BY THE TERMS. By clicking “I ACCEPT” or otherwise accessing or using any Services you also represent that you have the legal authority to accept the Terms on behalf of yourself and any party you represent in connection with your use of any Services. If you do not agree to the Terms, you are not authorized to use any Services. If you are an individual who is entering into these Terms on behalf of an entity, you represent and warrant that you have the power to bind that entity, and you hereby agree on that entity’s behalf to be bound by these Terms, with the terms “you,” and “your” applying to you, that entity, and other users accessing the Services on behalf of that entity.
+
+Human-readable summary of Sec 2: Please read these terms and only use our sites and services if you agree to them.
+3. Changes to the Terms
+
[…]
