### 2022-04-06 630ac4d
Update license.md
