### 2022-06-14 #170 @Alberto-GS
> @openjournals/dev — why is the editorial bot not obeying me here?

Hi, is there anything I can do?
