### 2022-06-14 #170 @arfon
@labarba – this assignment is failing as it doesn't look like @lorenanicole's profile was correctly set up on the JOSE site. @lorenanicole – can you make sure you've logged in with your ORCID here: https://jose.theoj.org and I can fix this.
