### 2022-08-05 #170 @arfon
> @labarba – this assignment is failing as it doesn't look like @lorenanicole's profile was correctly set up on the JOSE site. @lorenanicole – can you make sure you've logged in with your ORCID here: https://jose.theoj.org/ and I can fix this.

@labarba – we're still in the same situation here. @lorenanicole - have you logged in to the JOSE site? If so, could you tell me what your ORCID is so I can fix your account?
