### 2022-10-23 #170 @labarba
Another possibility is to add the videos in an iframe within the markdown, rather than a code cell. Like this example:
https://so-cool.github.io/you-only-write-thrice/
