### 2022-11-15 #170 @PedroVelez
Dear @labarba and  @kls2177 , thanks a lot for the comments.
We will fix the videos and reply to the comments regarding  the text as soon as possible.
best
pedro
