### 2023-01-04 #170 @labarba
hi everyone! 👋
@kls2177 – I know this is your first submission as editor, so please do not hesitate to ask for help. We need to find two reviewers. You can start with the [list of volunteers](https://bit.ly/jose-reviewers), also look through published papers and see if we can recruit a JOSE author (to give back to the community!), or have a Googling about the 'net 🚀
