### 2023-01-13 #170 @labarba
Hi @kls2177 👋 — JOSE is behind in the Open Journals technology, and we're still not upgraded to `@editorialbot`. Summon our old bot `@whedon` instead!
