### 2023-01-23 #193 @whedon
[REVIEW]: The Argo Online School: An e-learning tool to get started with Argo
**Submitting author:** @Alberto-GS (<a href="http://orcid.org/0000-0001-5781-9330">Alberto González</a>)
**Repository:** https://github.com/euroargodev/argoonlineschool
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1

- [x] **Version:** Does the release version given match the repository release (1.0)?

- [x] **Version:** Does the release version given match the repository release (1.0)?
