### 2023-01-24 #193 @PedroVelez
Dear @kls2177 on October 23rd we got from @labarba the following comments:

"hi @Alberto-GS and @PedroVelez 👋
Thank you for your patience. I did not receive a reply from the editor we initially assigned to this submission, despite several emails and DMs. (People have generally been much harder to reach in the past 2.5 years, and one can never know what is affecting them.) The good news is that JOSE had a campaign to recruit new editors, and we have 7 editors joining the board as we speak! (They need to be onboarded, though.)

In the meantime, I had a fresh look at the submission, in more depth, and let me anticipate a few things.

About the videos: note that JOSE considers videos as supplementary material that we do not review (*). I did notice that you have the videos on Google Drive, and for me they did not play on JupyterBook, as you have embedded them. It needed a "pop-out" to Drive. Have you considered putting them in a YouTube channel? (Perhaps unlisted? Vimeo is also an option). Then using [from IPython.display import YouTubeVideo](https://ipython.readthedocs.io/en/stable/api/generated/IPython.display.html#IPython.display.YouTubeVideo) they will be playable in the Jupyter notebook (and JupyterBook?).
On your paper, a reader has to wonder what Argos is all the way to the "Statement of Need," so I think that first paragraph there should be moved to the top of the Introduction.
Regarding the audience, you mention high-school students, which called my attention. Is there any experience involving high-school students in this topic? Do you have any experiences with different student populations that you could tell us about?
(*) We do not review videos, or slide presentations, or other forms of OCW, because they cannot be modified or easily reused. JOSE reviews open-source material that is reusable and allows derived works. Also, when you make an archive of the repository (source files for the JupyterBook), the videos will not be included."

we are working in these comments and shortly we will reply to them in this issue.
