### 2023-01-24 #193 @PedroVelez
Hi,
regarding  the first topic (1), we have tested versions using YouTube (see here https://www.euro-argo.eu/argo-online-school/Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.html ), however we do not like this versione since YouTube will suggest new videos that distract the 'reader' from the school as was as I know there is not way either using IPython or an iframe within the markdown.
We have no tested Vimeo but is similar since it promotes its own videos.
In general the Lessons1 and Lesson2 are meant to be seen in the main web page https://euroargodev.github.io/argoonlineschool/intro.html and Lesson 3 is the one to be downloaded to work with the Jupyter notebooks.

Please @kls2177  if you have another suggestion we will happy to test it.
best
Pedro
