### 2023-01-24 #193 @Alberto-GS
Hi all,

About (2), the "Statement of Need" has been placed where suggested.

About (3), we don't have previous experiences but this audience escope was proposed because here (at least in Spain) highschool students and first grade college students already have coding notions. We sadly don't have feedback with different student populations:( Do you find this scope audience too optimistic? Do you think it should be modified? @kls2177 @labarba

* I just realized there is an error on our affilitaion (probably my bad). It should be **Centro Oceanográfico de Canarias, Instituto Español de Oceanografía (IEO - CSIC**

Thanks!
