### 2023-01-26 #193 @kls2177
Hi @Alberto-GS  and @PedroVelez,

Thank you for your patience. Reviewers have now been assigned.

Regarding the videos, I am able to view them and I think using Youtube is the easiest way to proceed. However, please keep in mind "We do not review videos, or slide presentations, or other forms of OCW, because they cannot be modified or easily reused. JOSE reviews open-source material that is reusable and allows derived works. Also, when you make an archive of the repository (source files for the JupyterBook), the videos will not be included."

Regarding your audience, I suggest that you focus on the audience that this material was originally intended for rather than speculating as to who it could potentially be for.

I hope this helps!
