+++ .github/workflows/book.yml
@@ -1,34 +1,0 @@
-name: deploy-book
-
-# Only run this when the master branch changes
-on:
-  push:
-    branches:
-    - master
-
-# This job installs dependencies, build the book, and pushes it to `gh-pages`
-jobs:
-  deploy-book:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-
-    # Install dependencies
-    - name: Set up Python 3.7
-      uses: actions/setup-python@v1
-      with:
-        python-version: 3.7
-
-    - name: Install dependencies
-      run: |
-        pip install -r requirements.txt
-    # Build the book
-    - name: Build the book
-      run: |
-        jupyter-book build .
-    # Push the book's HTML to github-pages
-    - name: GitHub Pages action
-      uses: peaceiris/actions-gh-pages@v3.5.9
-      with:
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_dir: ./_build/html
