### 2023-05-08 #193 @francispoulin
I found a problem with one of the notebooks. I am going to post it here but if it's better to put it elsewhere, please let me know.

In L03-Chapter11_TheNetCDFFFormat, In [26], I get an error with the line `ax.plot(dataOPsubsampled.sst)`
