### 2023-05-08 #193 @francispoulin
In L03-Chapter22_ArgoDatabyFloat_Prof, in In [3]: the line should read as follows, otherwise the Data cannot be found and an error occurs,

`prof  = xr.open_dataset('../../Data/6901254/6901254_prof.nc')`.
