### 2023-05-08 #193 @francispoulin
In L03-Chapter31_ArgoDatabyData_Day, in In [3]: the first line should reference 2019, as that is what is found in the Data folder.

`dayADS = xr.open_dataset('../../Data/atlantic_ocean/2020/11/20201111_prof.nc')`.
