### 2023-05-08 #193 @francispoulin
In L03-Chapter41:

In [4] should refer to 2019 not 2020.

In[ 13] has an error as `another exception occured:`.
