### 2023-05-08 #193 @francispoulin
Ln L03-Chapter42:

In [3]: has an error `another exception occurred:`.
