### 2023-05-08 #193 @francispoulin
This is an interesting contribution to the field that will be useful to many young researchers.  I have pointed out a number of bugs that I found in the notebooks, but I presume these are easy to fix.

One general concern is that the first two lessons only involve watching videos and no actual python code. If that’s the case it’s nto clear the utility of including them in a notebook. I am okay with this but I feel like if there was some python in the first two lessons that might be better balanced.

Below are a few minor points that should be addressed.

    • The version should be 1.0 but the tag shows it to be v0.1.0.
    • The name is sometimes abreviated as AOS and other times AoS.  Either is fine but be consistant.
    • On the main documentation page, when I click on 1. The Argo Program, I get a 404 error.
    • The paper.md file does not have any references, however I do see them in the PDF.  I presume this is okay then but wanted to confirm before I click the box.
    • You should point out this is useful to high school, undergraduate and graduate students.
