### 2023-05-15 euroargodev/argoonlineschool#10 @djmallum
JOSE Review: Typos - djmallum
**L01**

C10

- Right-hand side of the page says “All the videos from the Argo Online Shcool are in YouTube” —> “All the videos from the Argo Online School are in YouTube”

C16

- “… its parameters are locally turned to reflect …” —> “… its parameters are locally tuned to reflect …”

**L02**

C10

- Under “The basis of the Argo Data System” it says “As explained in the videos of Lesson 2…” —> “As explained in the videos of Lesson 1”.

C12

- “… it is possible to efficiently collect data from? the main physical variables…” Not sure how this sentence is suppose to be. Maybe just an accidental ‘?’

C14

- “… the sensor itself. commas or etc: In both cases, …” Period should be a comma maybe?

**L03**

C10

- “… in a web-page, build using JubyterBook …” —> “… in a web-page, built using JupyterBook…”
- “…In the case that you …” —> “… In case that you…”

C11

- “ … of accessing the metadata stored in te Dataset” —> “… of accessing the metadata stored in the Dataset”
- “ print('We are exlploring', xrDS.attrs['dataset_title'] ...” --> “print('We are exploring', xrDS.attrs['dataset_title'] ...”
- Above typo repeats on the next line

C21

- “Check what useful is the Self-Describing property of the netCDF format.” Does not make sense and must be reworded.
- “We note that for for the first cycle there are two profiles (N_PROF=2) and 95 vertical levels… lets plot them:” Looks like there are 97 levels.

C23

- Typo in chapter title “trayectory” —> “trajectory”

C41

- “If we want to get a better map, we need cartopy, althuogh we could skip this step and”. Extra ‘and’ at the end of the sentence and typo in ‘although’

**Paper**

Notes on instructional design

- "the hands-on section of the AoS are also provide, whether the users” —> “the hands-on section of the AoS are also provided, whether the users”

Also note several instances of using AOS instead of AoS, as Francis mentioned.
