### 2023-05-15 euroargodev/argoonlineschool#11 @djmallum
JOSE Review: Lesson 1 - djmallum
CXX refers to chapter number in Lesson 1 only.

C10
- First video appears three times
- Link to learn more about satellite Jason would be useful

C13
- Link to learn more about satellite Jason would be useful

C14
- How are floats disposed after the battery expires? I expect this would be a pressing question for learners
- The word orbit under “Data Transmission System” should not be in quotes

C15
- Two videos and no guiding text is not ideal
