### 2023-05-15 euroargodev/argoonlineschool#14 @djmallum
JOSE Review: Checklist Comments - djmallum
**Documentation**

•	_Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?_

This needs to be added somewhere to the documentation. I don’t think it needs to be a part of the module, perhaps this could go in the paper but I’m not familiar with the standards of what to include and what to not include.

•	_Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support._

Likewise, I don’t see any guidelines anywhere. Both of these problems could go in the read me or the paper.

**Pedagogy**

•	_Pedagogy: Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)_

Lesson 3 has a few places where there is extraneous information. However, I don’t think it is necessary to remove anything but instead explain to the learner what they can safely ignore for this lesson when you run a command that lists all the information about an object. Further, I will also check this box if some guiding and explanatory sentences are added to lessons 1 and 2 in parts with very little text. You don’t want to repeat the video but instead just let the learner know what each chapter is for if the title is not descriptive.

**JOSE Paper**

Both of the following are not in the paper:

•	_Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?
•	Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?_

Please add both to the paper in some capacity.
