### 2023-05-19 #193 @kls2177
@Alberto-GS and @PedroVelez, the reviewers have completed their reviews and have comments for your consideration. I recommend that these comments be addressed before publication.

Thanks @francispoulin and @djmallum for your helpful comments. @djmallum, quick question - I do not see an issue on the argoonlineschool repo entitled "General Comments". Not sure if this got included elsewhere.

@francispoulin and @djmallum, would you be willing to review the revised version when available?

Just a few of my own comments: I agree with @francispoulin, that it would be nice to integrate some python examples earlier in the course. Perhaps, just some simple plotting examples - rather that providing some static images, ask the students to generate the plots using python. This would allow students who are learning python for the first time to get some gradual exposure before Lesson 3.

You mention that this material is for high school or graduate students - what about undergraduate students?

I do not necessarily agree with your statement that no pre-requisite knowledge is required. There are certainly expectations about general science and mathematics knowledge. I suggest providing some information about pre-requisite knowledge on the landing page of the Argo Online School website. For example, what pre-requisite oceanography, maths, chemistry, geography knowledge is expected of students? This is likely more relevant for high school students than post-secondary students. At the beginning of Lesson 3, you provide students with a link to the Earth and Environmental Data Science course to learn python if they need to, but it would be nice to provide students with information about these pre-requisite coding skills and resources at the beginning of the course.

I really like the interactive quizzes! These look great and give timely feedback to students. I would recommend integrating a few more of these at regular intervals to keep the students engaged. Structurally, I would also suggest embedding the quizzes into each lesson, rather than having them as a separate section at the end. I feel that this would better reflect the sequencing of the material.

Thank you for your patience! Please let me know if you have any questions.
