### 2023-05-19 euroargodev/argoonlineschool#15 @djmallum
JOSE Review: General Comments - djmallum
- Video thumbnails could be better. It should be possible on YouTube to choose the thumbnail so that prior to pressing play the video gives the AOS logo + a short title, rather being a frame of a person talking midsentence. Some of the videos you may not control so I think it is OK to not worry about the thumbnail for those videos.

- The intro describes the content well, but it should also include a component describing how the user should navigate the content. Reminding them that lesson 1 and 2 are very different from lesson 3.

- A time estimate is needed. This would be useful for learners and anyone hoping to adapt the module. The estimate should assume that the learner has all of the basic background.

- Quizzes should be placed at the end of each lesson, so the learners naturally go through the quizzes before moving on.

- Lesson 3 starts off too complex and relies heavily on assuming the learner knows a lot of python and is familiar with xarray. The biggest issue I have is that there are several instances in this lesson where there is some kind of text based output and no explanation of what they see is given. The text should also make an effort to explain what parts of these outputs can be safely ignored and which parts should be emphasized.

- The change in pace and complexity should also be mentioned in the intro.

- It is also be worth mentioning that Lesson 1-2 is itself a closed module and could be used by a learner or teacher without using lesson 3, assuming the goal is simply to learn about AGRO. This could be included in a new section in the paper where you describe how someone else could adapt the module.
