### 2023-07-24 be655f2
Typos in paper and references
