+++ paper.md
@@ -22,17 +22,17 @@ bibliography: paper.bib
 
 # Summary
 
-The **Argo Online School (AoS)** is a set of videos, animations and hands-on python driven jupyter notebooks designed to make the data, of the more than 4000 profiling floats that constitute the Argo program, accessible. The Argo program samples, in near real time, the upper 2000 metres of the ocean using a fleet of floats that drift with the ocean currents. The AoS is integrated by 25 sections about the Argo program and its data, organized into a total of 3  lessons and 2 self-assessment sections:
+The **Argo Online School (AOS)** is a set of videos, animations and hands-on python driven jupyter notebooks designed to make the data, of the more than 4000 profiling floats that constitute the Argo program, accessible. The Argo program samples, in near real time, the upper 2000 metres of the ocean using a fleet of floats that drift with the ocean currents. The AOS is integrated by 25 sections about the Argo program and its data, organized into a total of 3  lessons and 2 self-assessment sections:
 
 * Lesson 1 is intended to introduce the basics of Argo, its objectives and key elements, such as the structure of the Argo floats and their operation in the open ocean.
 
 * Lesson 2 focuses on the data coming from the Argo network, from its organization to its accessibility. Data quality control is also addressed through its two main modes: Real-Time and Delayed-Mode.
 
-* Lesson 3 is the AoS hands-on component. It offers a set of instructions to prepare a python environment in case the user wants to run the python Jupyter Notebooks in a local machine. This environment already includes recommended packages. The walkthrough of Lesson 3 shows how to work with the netCDF format, how to access and process Argo data, and how to produce fancy plots for understanding information from Argo data.
+* Lesson 3 is the AOS hands-on component. It offers a set of instructions to prepare a python environment in case the user wants to run the python Jupyter Notebooks in a local machine. This environment already includes recommended packages. The walkthrough of Lesson 3 shows how to work with the netCDF format, how to access and process Argo data, and how to produce fancy plots for understanding information from Argo data.
 
-The target audience of the AoS are high school or early graduate students, and the programming content included in Lesson 3 provides an ideal opportunity to support students with technical or science curriculum. It is recommended not to skip any lesson in the AoS as the content is carefully connected from less to greater complexity, guaranteeing a progressive learning. The user has the opportunity to self-assess the learning process carried out during their course in the AoS through the proposed quizzes.
+The target audience of the AOS are high school or early graduate students, and the programming content included in Lesson 3 provides an ideal opportunity to support students with technical or science curriculum. It is recommended not to skip any lesson in the AOS as the content is carefully connected from less to greater complexity, guaranteeing a progressive learning. The user has the opportunity to self-assess the learning process carried out during their course in the AOS through the proposed quizzes.
 
-The AoS has been designed to have the possibility to be expanded to follow the implementation of new features in the Argo program.
+The AOS has been designed to have the possibility to be expanded to follow the implementation of new features in the Argo program.
 
 # Statement of need
 
@@ -40,15 +40,15 @@ Argo [@roemmich-2019; @riser-2016] is an international program that collects inf
 
[…]
