### 2023-07-25 e1a865b
Create contribute_report_support.md
This is an provided answer based on  JOSE Review: Checklist Comments - djmallum #14 (MAY 15)

• Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support.

Likewise, I don’t see any guidelines anywhere. Both of these problems could go in the read me or the paper.
