+++ _toc.yml
@@ -11,6 +11,8 @@ parts:
     - file: Lessons/L01_TheArgoProgram/Chapter15_OceanographicProfile.ipynb
     - file: Lessons/L01_TheArgoProgram/Chapter16_TechInnovations.ipynb
     - file: Lessons/L01_TheArgoProgram/Chapter17_RecapLesson1.ipynb
+  chapters:
+  - file: Lessons/Quizzes/Quiz10
 
 - caption: Lesson 2
   chapters:
@@ -21,7 +23,9 @@ parts:
     - file: Lessons/L02_TheArgoData/Chapter14_RealTimeDelayedMode.ipynb
     - file: Lessons/L02_TheArgoData/Chapter15_QualityFlagMean.ipynb
     - file: Lessons/L02_TheArgoData/Chapter16_GetAccesOrganized.ipynb
-    - file: Lessons/L02_TheArgoData/Chapter17_RecapLesson2.ipynb    
+    - file: Lessons/L02_TheArgoData/Chapter17_RecapLesson2.ipynb 
+  chapters:
+  - file: Lessons/Quizzes/Quiz10
 
 - caption: Lesson 3
   chapters:
