+++ _toc.yml
@@ -11,7 +11,6 @@ parts:
     - file: Lessons/L01_TheArgoProgram/Chapter15_OceanographicProfile.ipynb
     - file: Lessons/L01_TheArgoProgram/Chapter16_TechInnovations.ipynb
     - file: Lessons/L01_TheArgoProgram/Chapter17_RecapLesson1.ipynb
-    - file: Lessons/Quizzes/Quiz10
 
 - caption: Lesson 2
   chapters:
@@ -23,7 +22,6 @@ parts:
     - file: Lessons/L02_TheArgoData/Chapter15_QualityFlagMean.ipynb
     - file: Lessons/L02_TheArgoData/Chapter16_GetAccesOrganized.ipynb
     - file: Lessons/L02_TheArgoData/Chapter17_RecapLesson2.ipynb 
-    - file: Lessons/Quizzes/Quiz10
 
 - caption: Lesson 3
   chapters:
