+++ _toc.yml
@@ -52,8 +52,7 @@ parts:
     
 - caption: Quizzes
   chapters:
-  - file: Lessons/Quizzes/Quiz10
-  - file: Lessons/Quizzes/Quiz20
+
 
 
     
