### 2024-07-05 #193 @Alberto-GS
Hi @kls2177 I'm trying to resolve some serious bugs but I'm having trouble pushing all the commits https://github.com/euroargodev/argoonlineschool/actions Could you shed some light on the problem? I'm familiar with GitHub but I'm not very skilled. I've installed all the dependencies, packages and updates on my local computer and it seems to work just fine. In fact, I can see the changes correctly. Once I push it to the repository, I get this GitHub report containing errors. I'm guessing something is going on with HDF5 but I still don't know how to solve it after several hours of searching (it's probably just a silly thing). Thanks.
