+++ .github/workflows/book.yml
@@ -21,7 +21,7 @@ jobs:
 
     - name: Install dependencies
       run: |
-        pip install -r requirements.txt
+        pip install -r requirementsJL.txt
     # Build the book
     - name: Build the book
       run: |
