+++ requirementsJL.txt
@@ -1,0 +1,7 @@
+jupyter-book>=0.7.0b
+argopy
+scipy>=1.1.0
+xarray>=0.14
+scikit-learn>=0.21
+ipython>=5.0.0
+numpydoc>=0.6.0
