+++ .github/workflows/book.yml
@@ -21,6 +21,7 @@ jobs:
 
     - name: Install dependencies
       run: |
+        pip install --upgrade pip
         pip install -r requirementsJL.txt
     # Build the book
     - name: Build the book
