+++ _sources/intro.md
@@ -13,7 +13,7 @@ Given the amount of data gathered by the Argo network of floats, the associated 
 
 The AOS is organized into three main sections:
 
-**1. [The Argo Program](https://euroargodev.github.io/argoonlineschool/Lectures/L01_TheArgoProgram/C10_TheArgoProgram_intro.html)**
+**1. [The Argo Program](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.html)**
 
 In this section and using a set of videos you will learn about:
 * [What is the Argo network?](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter12_WhatArgoNetwork.html)
@@ -24,17 +24,18 @@ In this section and using a set of videos you will learn about:
 * [Recap Lesson 1](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter17_RecapLesson1.html)
 
 
-**2. [The Argo Data](https://euroargodev.github.io/argoonlineschool/Lectures/L02_TheArgoData/C10_TheArgoData_intro.html)**
+**2. [The Argo Data](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter10_TheArgoData_intro.html)**
 
 In this section and using a set of videos you will learn about:
 * [The Argo General Data Stream](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter12_ArgoDataStream.html)
 * [What are the Argo data and its format?](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter13_WhatArgoDataFormat.html)
-* [Real-Time and Delayed mode data.](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter14_RealTimeDelayedMode.html)
-* [Where can I get access to data and how is it organized?](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter15_GetAccesOrganized.html)
+* [Real-Time and Delayed mode data](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter14_RealTimeDelayedMode.html)
+* [What does a quality control flag mean?](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter15_QualityFlagMean.html)
+* [Where can I get access to data and how is it organized?](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter16_GetAccesOrganized.html) 
 * [Recap Lesson 2](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoData/Chapter17_RecapLesson2.html)
 
 
-**3. [Using the Argo data](https://euroargodev.github.io/argoonlineschool/Lectures/L03_UsingArgoData/C10_UsingArgoData_intro.html)**
+**3. [Using the Argo data](https://euroargodev.github.io/argoonlineschool/Lessons/L03_UsingArgoData/Chapter10_UsingArgoData_intro.html)**
 
 In this section and using hands-on jupyter notebooks you will learn about:
 * [The netCDF format](https://euroargodev.github.io/argoonlineschool/Lessons/L03_UsingArgoData/Chapter11_TheNetCDFFormat.html)
@@ -43,7 +44,7 @@ In this section and using hands-on jupyter notebooks you will learn about:
[…]
