### 2024-08-08 #193 @Alberto-GS
Hello @kls2177, after some checks @PedroVelez solved the bug. Apparently there was some issue between the workflow and the netCDF package. Now we're on the road again.
