### 2024-08-08 c7ca054
fixing issue JOSE Review: Typos - djmallum #10
