### 2024-08-08 euroargodev/argoonlineschool#10 @PedroVelez
Hi, thanks!  we have fixed all the typos listed in here.
