### 2024-08-08 euroargodev/argoonlineschool#11 @PedroVelez
Hi, thanks! we have fixed all the typos listed in here and the suggestions.
