+++ _toc.yml
@@ -11,6 +11,7 @@ parts:
     - file: Lessons/L01_TheArgoProgram/Chapter15_OceanographicProfile.ipynb
     - file: Lessons/L01_TheArgoProgram/Chapter16_TechInnovations.ipynb
     - file: Lessons/L01_TheArgoProgram/Chapter17_RecapLesson1.ipynb
+    - file: POcultarVideos.ipynb
     - file: Lessons/Quizzes/Quiz10.ipynb      
     
 - caption: Lesson 2
