+++ POcultarVideos.md
@@ -1,0 +1,9 @@
+## Pruebas de como poner videos
+
+### Using markdow - Opcion 2 - La prueba definitiva
+
+In this video we .... 
+
+<center>
+<iframe width="560" height="315" src="https://www.youtube.com/embed/88XFuf6w2v0?si=Jir22e7h-fL9aBg1" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+</center>
