+++ Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.md
@@ -1,0 +1,23 @@
+# The Argo Program
+
+<img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/logoArgo.png" alt="Argo status" width="100"/>
+
+###  Welcome to the Argo Online School
+<center>
+<iframe width="560" height="315" src="https://www.youtube.com/embed/-TE64w4hi2o?si=fsl8viQlDz2QIsxk&amp;controls=1" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+</center>
+
+Argo is a large-scale global set of profiling floats that measure temperature and salinity versus pressure, in addition to six [biogeochemical variables](https://biogeochemical-argo.org/): oxygen concentration, nitrate concentration, pH, chlorophyll-a concentration, suspended particles and downwelling irradiance. Argo is the main component of the ocean observing system. This is the current distribution of the Argo matrix.
+
+<center>
+    <img src="http://sio-argo.ucsd.edu/statusbig.gif" alt="Argo status" width="800"/>
+</center>
+
+
+<center>
+<iframe width="560" height="315" src="https://drive.google.com/file/d/1PvpDuyAdDX0CidiKRZUhIAQSJJDzRdLc/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+</center>
+
+<center>
+<iframe width="560" height="315" src="https://drive.google.com/file/d/16dRAJAU9_6arlkF2Fzw59B-E-hua_qOk/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+</center>
