+++ _sources/Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.md
@@ -2,22 +2,28 @@
 
 <img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/logoArgo.png" alt="Argo status" width="100"/>
 
-###  Welcome to the Argo Online School
-<center>
-<iframe width="560" height="315" src="https://www.youtube.com/embed/-TE64w4hi2o?si=fsl8viQlDz2QIsxk&amp;controls=1" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
-</center>
+####  Welcome to the Argo Online School
 
-Argo is a large-scale global set of profiling floats that measure temperature and salinity versus pressure, in addition to six [biogeochemical variables](https://biogeochemical-argo.org/): oxygen concentration, nitrate concentration, pH, chlorophyll-a concentration, suspended particles and downwelling irradiance. Argo is the main component of the ocean observing system. This is the current distribution of the Argo matrix.
+&nbsp;&nbsp;
+<center><iframe width="560" height="315" src="https://www.youtube.com/embed/-TE64w4hi2o?si=fsl8viQlDz2QIsxk&amp;controls=1" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+&nbsp;&nbsp;
 
-<center>
-    <img src="http://sio-argo.ucsd.edu/statusbig.gif" alt="Argo status" width="800"/>
-</center>
+Argo is a large-scale global set of profiling floats that measure temperature and salinity versus pressure, in addition to six [biogeochemical variables](https://biogeochemical-argo.org/): oxygen concentration, nitrate concentration, pH, chlorophyll-a concentration, suspended particles and downwelling irradiance. Argo is the main component of the ocean observing system. 
 
+This is the actual distribution of the Argo fleet of floats.
 
-<center>
-<iframe width="560" height="315" src="https://drive.google.com/file/d/1PvpDuyAdDX0CidiKRZUhIAQSJJDzRdLc/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
-</center>
+&nbsp;
 
-<center>
-<iframe width="560" height="315" src="https://drive.google.com/file/d/16dRAJAU9_6arlkF2Fzw59B-E-hua_qOk/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
-</center>
+<center><img src="http://sio-argo.ucsd.edu/statusbig.gif" alt="Argo status" width="800"/></center>
+
+####  Why study the ocean
+
+Let’s explain briefly in the following wideo why we need to study the ocean. 
+
+&nbsp;&nbsp;<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1PvpDuyAdDX0CidiKRZUhIAQSJJDzRdLc/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>&nbsp;&nbsp;
+
+####  A new way to observe the ocean
+
[…]
