+++ _sources/Lessons/L01_TheArgoProgram/Chapter12_WhatArgoNetwork.md
@@ -1,0 +1,23 @@
+## What is the Argo network?
+
+### Argo was born to combat the lack of ocean data
+
+In 1999, to combat this lack of data, an innovative step was taken by scientists to greatly improve the collection of observations inside the ocean through the increased sampling of old and new quantities and increased coverage in terms of time and area. That step was Argo.
+
+&nbsp;&nbsp;<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1OyrQ9ZIbZc-YjRXYaKmqer6baC09oWJu/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+
+### Why is called Argo? 
+
+&nbsp;&nbsp;<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1TdBc4tQumAzPuHKXC1yLdZkdDatNcB1e/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+
+### A fundamental piece of Global Ocean Observation System
+
+&nbsp;&nbsp;<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1KTVaDgugqksmY14Pe6vV1Kgo_W2MvViZ/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+
+&nbsp;&nbsp;
+
+The Argo Program is managed by teams of scientists and data experts who ensure that the program is run as efficiently and effectively as possible and that standards are maintained at the highest possible level. The total annual cost of Argo is estimated at $40million each year.
+
+The observations are used in a wide range of applications that focus on examining the relevant weather variability, from seasonal to decadal and multidecadal timescales, climate change, the improvement of models for the interactions between the ocean and the atmosphere and prediction systems. In addition, the Argo floats deliver quality-controlled data and products to researchers in the fields of climate and oceanography, and also to the operational communities, such as the Copernicus Marine Monitoring Service (CMEMS). 
+
+Behind the Argo program, there is an array of management and coordination agencies of different types, whose existence has guaranteed the success of the program, from the initial stages to the present. More information about these will be provided in [Lesson 2- The Argo data](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter10_TheArgoData_intro.html#).
