### 2024-08-12 8af222d
Creating text for each video in L01
