+++ Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.md
@@ -4,11 +4,13 @@
 
 ####  Welcome to the Argo Online School
 
+The goal of the Argo Online school ([AoS](https://euroargodev.github.io/argoonlineschool/)) is to teach the basic foundations to use and understand the Argo data. Wellcome!
+
 &nbsp;&nbsp;
-<center><iframe width="560" height="315" src="https://www.youtube.com/embed/-TE64w4hi2o?si=fsl8viQlDz2QIsxk&amp;controls=1" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+<center><iframe width="800" height="450" src="https://www.youtube.com/embed/DYiJ3Od2coI?si=yWyEPv6sXKc4xdqF" title="Argo Online School 101" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
 &nbsp;&nbsp;
 
-Argo is a large-scale global set of profiling floats that measure temperature and salinity versus pressure, in addition to six [biogeochemical variables](https://biogeochemical-argo.org/): oxygen concentration, nitrate concentration, pH, chlorophyll-a concentration, suspended particles and downwelling irradiance. Argo is the main component of the ocean observing system. 
+Argo is a large-scale global fleet of profiling floats that measure temperature and salinity versus pressure, in addition to six [biogeochemical variables](https://biogeochemical-argo.org/): oxygen concentration, nitrate concentration, pH, chlorophyll-a concentration, suspended particles and downwelling irradiance. Argo is the main component of the ocean observing system. 
 
 This is the actual distribution of the Argo fleet of floats.
 
@@ -20,10 +22,10 @@ This is the actual distribution of the Argo fleet of floats.
 
 Let’s explain briefly in the following wideo why we need to study the ocean. 
 
-&nbsp;&nbsp;<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1PvpDuyAdDX0CidiKRZUhIAQSJJDzRdLc/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>&nbsp;&nbsp;
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/DYiJ3Od2coI?si=odUzAzo0m0gTSVZX&amp;start=1" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>&nbsp;&nbsp;
 
 ####  A new way to observe the ocean
 
 Marine environment data is a valuable asset. Quick access to reliable and accurate information is vital to address threats to the marine environment, in the development of policies and laws to protect vulnerable areas of our coasts and oceans, to understand trends, and to forecast future changes. Unfortunately, the collection, storage and access to marine data has been fragmented for many years.
 
[…]
