### 2024-08-12 euroargodev/argoonlineschool#12 @PedroVelez
> C10
>
> * I don’t understand the section titled “The WMO: a unique identifier for each float”. The identified that each float carries (which I assume is different for each float) is called the WMO? The same name as the organization?
> * When you say “As explained in the videos of lesson 2” do you mention the current lesson? If so, say “this lesson” instead. If not, perhaps it is a typo.
> * NetCDF files are brought up in this lesson. Remind them that lesson 3 goes into it in more detail.

Thanks a lot, we have udpated the chapter to take into account this suggestions.

> C12
>
> * “The basis of the Argo Data System” has some repeated text at the start which is also found in C10.
> * Below the infographic there are sections a - e. Are these connected to the infographic? They appear to be loosely related. I would suggest connecting the list directly to the graphic if possible.
>
> C14
>
> * In the paragraph describing the test examples, ensure your test numbers are correct. For example, ‘deepest pressure test’ is test 19 and appears to be a more appropriate name for test 1. I could be mistaken here, but worth double checking.
> * The excellent graph at the beginning of the section showing some examples of what bad data can look like is a bit misplaced and largely goes unexplained. It also appears after the paragraph where DeltaS is defined. It should come after or right before.
> * The paragraph that starts with ‘Biodeposition’ starts with a few sentences which are stated exactly in the video. I don’t see a good reason to repeat yourself here. I suggest starting this paragraph with “Biodeposition mainly affects the conductivity sensors that allow …”
> * In a bullet point later on this page there is some extra text in brackets. I am unsure whether it needs to be included or not.
> * The graph at the end of this page is great but is missing a horizontal axis label and a readable legend. The text below the graph serves as a legend, so perhaps you could remove the legend entirely and let the text do the describing. I also recommend moving the text above the graph so as to avoid confusing the learner.
>
> C16
>
> * The last three videos on this page are very informative. However, they appear to use a webpage that is not linked to anywhere in the AOS school. I recommend proceeding these videos with a few sentences describing the following videos and a short description of how to access what the learner is seeing in the videos.
>
> Quiz
>
> * Question 5 should be “what is the purpose of The Argo Project Office” or something similar
