### 2024-08-14 euroargodev/argoonlineschool#12 @PedroVelez
I close the issue, since all the comment has been taking into account. We haver aso renamed some of the chapters and its numbering since we missed Ch11
