+++ Lessons/L02_TheArgoData/Chapter22_WhatArgoDataFormat.md
@@ -1,0 +1,6 @@
+#   What are the Argo data and its format?
+
+
+In general, when we talk about data, we mean a quantity that is measured. In this case, from the ocean. For example, the temperature at 4.2 meters below the surface is data. Isolated data itself does not constitute information. However, its processing is what provides us useful information for decision-making, reducing uncertainty, or performing calculations.
+
+&nbsp;&nbsp; <center><iframe width="800" height="450" src="https://drive.google.com/file/d/1wf2rftFTG6OrTdZTdc3UlUnW-6nhJYQH/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
