+++ _sources/Lessons/L02_TheArgoData/Chapter23_QualityControl.md
@@ -1,0 +1,78 @@
+# Quality control
+
+
+#### Exhaustibe quality controls are needed
+
+There are numerous reasons why data can be reported abnormally during the Argo profiler's operating cycles. It may be due to problems related to the CTD sensors, either external interference to the sensor, for instance, due to biodeposition, or intrinsic problems of the sensor itself. In both cases, incorrect data with spikes, biases, sensor drifts, etc., can be reported as bad data or possible bad data. Therefore Argo needs exhaustibe quality controls.
+
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/1PDwmj7AsbDoy9tLNMdtTc7q_Mey2LvxH/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>&nbsp;&nbsp;
+
+
+#### The firtst quality control: Real-Time 
+
+In real time, this is less than 24 hours since the data was measured by the Argo floats, a first Quality Control is carried out: the Real time Quality control (RTQC).
+
+<center><img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/GIF_RT_QualityControl.gif" alt="Argo status" width="800"/></center>
+
+Let's summarize this step:
+
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/18NM9vPb1gJ94keeXA50h6-T6zGvfoV6_/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>&nbsp;&nbsp;
+
+For instance, test 19 verifies whether the maximum pressure measured by a float is higher than the ocean depth in the location of the float; test 4 verifies if the location reported by the float is actually in the ocean, or if by error it is reporting a position overland; test 18 checks whether the float is sending, again and again, the same data, etc. We don’t expect you to learn what every test is about, but if you are curious about all of them, take a look at the [Argo Quality Control Manual for CTD and Trajectory Data](https://archimer.ifremer.fr/doc/00228/33951/), and here there is a list of the RTQC tests:
+
+<center><img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/QC_Test.jpg" alt="Argo status" width="800"/></center>
+
+*The full set of Argo Quality Control Manuals can be found here [ADMT Documentacion](http://www.argodatamgt.org/Documentation)*
+
+All the data will be distributed in the Global Telecomunication Systrem (GTS), with its corresponding quality flags, that give an indication of the quality of the data. On the other hand, the data and its flags will be also converted to the netCDF format and sent to the Argo Global Data Management Centers (GDAC).
+
+#### Problems in the sensors vs natural variability
+
+However, detecting problems related to the CTD sensors can sometimes be very difficult. Even more when we take into account the concept of “natural variability”. Natural variability is just the natural change of the properties of the ocean and it may happen in periods ranging from seconds to millennia. For instance, changes occur daily, such as the tides; sporadically, such as storm surges or tsunamis; and seasonally, such as the simple warming of coastal waters in summer. In addition to the uncertainty associated with sensor errors, natural systems have an uncertainty associated with the range of natural variability itself. In such cases, it is very difficult to distinguish between the natural variability observed and the errors.
+
+Since there is also natural variability in the temperature and salinity that can be mistaken for sensor malfunction, we have to be careful not to treat actual natural variability as sensor malfunction and correct it. 
+
[…]
