+++ Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.md
@@ -6,8 +6,7 @@
 
 The goal of the Argo Online school ([AoS](https://euroargodev.github.io/argoonlineschool/)) is to teach the basic foundations to use and understand the Argo data. Wellcome!
 
-&nbsp;&nbsp;
-<center><iframe width="800" height="450" src="https://www.youtube.com/embed/DYiJ3Od2coI?si=yWyEPv6sXKc4xdqF" title="Argo Online School 101" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/DYiJ3Od2coI?si=yWyEPv6sXKc4xdqF" title="Argo Online School 101" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
 &nbsp;&nbsp;
 
 Argo is a large-scale global fleet of profiling floats that measure temperature and salinity versus pressure, in addition to six [biogeochemical variables](https://biogeochemical-argo.org/): oxygen concentration, nitrate concentration, pH, chlorophyll-a concentration, suspended particles and downwelling irradiance. Argo is the main component of the ocean observing system. 
