+++ Lessons/L01_TheArgoProgram/Chapter15_TechInnovations.md
@@ -1,0 +1,39 @@
+## Technological innovations
+
+Argo is constantly trying to improve the quality and coverage of its data, which means that some technological challenges need to be addressed in order to have floats operate well in all parts of the global ocean. Remember that almost all floats are deployed into the ocean and not retrieved and thus the floats must be robust, long-lasting and accurate in a sometimes very harsh marine environment. Read through the different innovations below that have helped Argo improve and expand its coverage.
+
+
+<center>
+<iframe width="560" height="315" src="https://drive.google.com/file/d/1A043Rlbuj7SXiRFz7V0yCbr3hnjZWyTf/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+</center>
+
+
+#### Improving float lifetime
+
+If floats live for a long time, more profiles can be delivered and fewer floats need to be deployed each year. This helps to keep the cost of the Argo Program from growing exponentially each year and reduces the pollution associated with float deployment.  One of the most basic ways of extending float lifetimes is making sure that there is enough battery life inside the floats to power them for years.  When Argo began in the early 2000s, all floats used alkaline batteries and had a goal float lifetime between  3 – 5 years.  After monitoring technical data sent by floats, it became clear that after about three years, the alkaline batteries ran out of power. A few alternatives were tested to improve the battery performance and now most Argo floats are powered by lithium or hybrid batteries. These batteries can cope better with colder conditions and also handle large current demands, like those needed to start the float ascent or descent. In addition, more battery packs were added to some float types that use lots of energy, like kinds of floats that carry 4 to 6 additional sensors.
+
+#### High-speed communications
+
+At the beginning of Argo, floats were equipped with transmitters that had low bandwidth and required the floats to be on the surface for many hours in order to upload all their data (up to 18 hours in the tropics). An unfortunate side effect of this in areas that had large surface currents was that the floats could be swept ashore or over areas that were shallower than their 2000m profile depth. This limited the locations that floats could be deployed in and especially made it hard to operate in marginal seas, which often have shallow areas.  However, communications have now improved and most floats use transmitters that are high bandwidth and only require the floats to be on the surface for 15–30 minutes. This has greatly reduced floats being washed ashore or lost to grounding (getting stuck on the bottom because the bottom depth is shallower than 2000m). Another benefit of high-speed communications is that much more data can be sent, allowing higher resolution profiles. In addition, float operators can send instructions to the floats to change their mission parameters, which helps to avoid hazards or sample upcoming weather events. However, high-speed communication systems do not offer high-quality positions, so floats now carry GPS devices whose acquired positions are more accurate than those obtained previously In turn, more accurate surface positions can lead to more accurate water velocity estimations.
+
+#### Ice detection
+
[…]
