+++ Lessons/L01_TheArgoProgram/Chapter12_NeedArgo.md
@@ -1,23 +1,35 @@
 ## Why do we need Argo and what are its objectives?
 
 
-<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1DfGg8uddYSKy2LS3ixlheihc3VNL32Ra/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+The lack of sustained observations of the atmosphere, oceans, and land have hindered the development and validation of climate models
 
-<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1pivKHZE5bpFZCRNe2bnv1S_XP6vXV2mH/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/_t3CTpDI3PA?si=ry47LKgkpSVk8A_0" title="Why do we need Argo and what are its objectives?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>&nbsp;&nbsp;
 
-Argo's general objectives are:
+
+#### Argo's general objectives :
 
 - To provide a quantitative description of the changing state of the ocean's surface layer and the patterns of sea climate variability from months to decades, including the storage and transportation of heat and freshwater.
 - Provide data that will also be used to initialize ocean prediction and the coupled atmosphere-ocean data assimilation models and model testing.
 - Document climate variability from seasons to decades, thus helping to understand its predictability.
 - Provide data collected or the calibration of satellite Jason's altimetric observations by means of measuring the subsurface temperature, salinity and speed of currents. Jason is a series of satellite that provide global sea surface height observations for climate monitoring and ocean and seasonal forecasts. More information could be found [here](https://www.eumetsat.int/our-satellites/jason-series)
 
-The data is provided free via the Global Data Assembly Centres (GDAC) located In Brest, France, and Monterey, California, after having gone through automated quality control (QC).
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/lpAE2WtMt-E?si=NgmGn3pBDWsLJyxk" title=" Argo's general objectives " frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>&nbsp;&nbsp;
 
-<center><iframe width="560" height="315" src="https://drive.google.com/file/d/170rjpguO0e8YDQcpyGgnU1kukpcRZTtx/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+The data is provided free via the Global Data Assembly Centres (GDAC) located In Brest, France, and Monterey, California, after having gone through automated quality control (QC).
 
 Argo's array growth is monitored by the Technical Coordinator at [OceanOPS](https://www.ocean-ops.org/board), which is located at the French Research Institute for Exploitation of the Sea ([IFREMER](https://wwz.ifremer.fr/)) in Brest, France.
 
[…]
