+++ _config.yml
@@ -1,7 +1,7 @@
 # Book settings
 title                       : Argo Online School
 author                      : Euro Argo
-copyright                   : 2020-2021
+copyright                   : 2020-2024
 logo                        : ./images/logoAoS.png
 exclude_patterns            : [Prueba*.ipynb, Info.ipynb, README.md]
 description : >-

[…]
