+++ _sources/Lessons/L01_TheArgoProgram/Chapter14_OceanographicProfile.md
@@ -1,0 +1,11 @@
+## What is an oceanographic profile?
+
+Each of the temperature and salinity data sets measured from 2000 meters depth to surface is called an oceanographic profile, which is why the Argo floats are also known as "oceanographic profilers".  In the following video you will find a detailed explanation and some examples:
+
+
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/m2hj1WrteCg?si=jPixZ_vXck0KAc3m&amp;start=6" title="What is an oceanographic profile?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+
+#### Argo has been revolutionary
+Argo represents a revolution in ocean observation, since it provides data, freely and free of charge, a rate 20 times higher compared to traditional methods on board oceanographic records. Here we present some more interesting facts:
+
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/HjOG0TgMVGo?si=UAmTlAanhANDvCG3&amp;start=3" title="Argo has been revolutionary" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
