+++ _sources/Lessons/L01_TheArgoProgram/Chapter11_WhatArgoNetwork.md
@@ -4,15 +4,15 @@
 
 In 1999, to combat this lack of data, an innovative step was taken by scientists to greatly improve the collection of observations inside the ocean through the increased sampling of old and new quantities and increased coverage in terms of time and area. That step was Argo.
 
-&nbsp;&nbsp;<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1OyrQ9ZIbZc-YjRXYaKmqer6baC09oWJu/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/4_G78ZLgjgU?si=HhwbD3TgbQ3m8nwJ" title="Argo was born to combat the lack of ocean data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
 
 ### Why is called Argo? 
 
-&nbsp;&nbsp;<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1TdBc4tQumAzPuHKXC1yLdZkdDatNcB1e/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/YBOo7McONUk?si=IgNWaZhP-ERceSku" title="Why is called Argo?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
 
 ### A fundamental piece of Global Ocean Observation System
 
-&nbsp;&nbsp;<center><iframe width="560" height="315" src="https://drive.google.com/file/d/1KTVaDgugqksmY14Pe6vV1Kgo_W2MvViZ/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/D6iAsqsdSx0?si=IF9QbGKfLPCWR3_n" title="A fundamental piece of Global Ocean Observation System" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
 
 &nbsp;&nbsp;
 
