+++ Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.md
@@ -6,7 +6,7 @@
 
 The goal of the Argo Online school ([AoS](https://euroargodev.github.io/argoonlineschool/)) is to teach the basic foundations to use and understand the Argo data. Wellcome!
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/DYiJ3Od2coI?si=yWyEPv6sXKc4xdqF" title="Argo Online School 101" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/DYiJ3Od2coI?si=yWyEPv6sXKc4xdqF" title="Welcome to the Argo Online School" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
 &nbsp;&nbsp;
 
 Argo is a large-scale global fleet of profiling floats that measure temperature and salinity versus pressure, in addition to six [biogeochemical variables](https://biogeochemical-argo.org/): oxygen concentration, nitrate concentration, pH, chlorophyll-a concentration, suspended particles and downwelling irradiance. Argo is the main component of the ocean observing system. 
@@ -21,10 +21,10 @@ This is the actual distribution of the Argo fleet of floats.
 
 Let’s explain briefly in the following wideo why we need to study the ocean. 
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/DYiJ3Od2coI?si=odUzAzo0m0gTSVZX&amp;start=1" title="Argo Online School 102" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>&nbsp;&nbsp;
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/DYiJ3Od2coI?si=odUzAzo0m0gTSVZX&amp;start=1" title="Why study the ocean" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>&nbsp;&nbsp;
 
 ####  A new way to observe the ocean
 
 Marine environment data is a valuable asset. Quick access to reliable and accurate information is vital to address threats to the marine environment, in the development of policies and laws to protect vulnerable areas of our coasts and oceans, to understand trends, and to forecast future changes. Unfortunately, the collection, storage and access to marine data has been fragmented for many years.
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/7P2Rqw52PXk?si=_awesB9L8Sh1S61y" title="Argo Online School 103" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
[…]
