### 2024-08-15 8ae7953
Videos from L02 - 234 in YT
