+++ Lessons/L02_TheArgoData/Chapter21_ArgoDataStream.md
@@ -27,4 +27,4 @@ The **Argo Information Center (AIC)** oversees the processes to ensure that all 
 The “final data pool” is complemented with the inclusion of a historical climatological database worldwide. This database contains all the profiles collected by the Argo profilers in delayed mode and all the hydrographic profiles of those collected in selected oceanographic surveys. 
 
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/1zRRrZD0QI41iFbkoOZnBqoaq2yPgEQID/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/hDhV80TB_QA?si=yEkEcxpBKwW-2KwA&amp;start=1" title="All the data is saved" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
