+++ Lessons/L02_TheArgoData/Chapter23_QualityControl.md
@@ -5,7 +5,9 @@
 
 There are numerous reasons why data can be reported abnormally during the Argo profiler's operating cycles. It may be due to problems related to the CTD sensors, either external interference to the sensor, for instance, due to biodeposition, or intrinsic problems of the sensor itself. In both cases, incorrect data with spikes, biases, sensor drifts, etc., can be reported as bad data or possible bad data. Therefore Argo needs exhaustibe quality controls.
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/1PDwmj7AsbDoy9tLNMdtTc7q_Mey2LvxH/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>&nbsp;&nbsp;
+&nbsp;&nbsp;<center>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/pvJN-6OXtDw?si=0Br4TKIy4JNEkI-R&amp;start=1" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+</center>&nbsp;&nbsp;
 
 
 #### The firtst quality control: Real-Time 
@@ -16,7 +18,7 @@ In real time, this is less than 24 hours since the data was measured by the Argo
 
 Let's summarize this step:
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/18NM9vPb1gJ94keeXA50h6-T6zGvfoV6_/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>&nbsp;&nbsp;
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/6GHUPfFHSNo?si=1fZfiZMBnJiQPYu0" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>&nbsp;&nbsp;
 
 For instance, test 19 verifies whether the maximum pressure measured by a float is higher than the ocean depth in the location of the float; test 4 verifies if the location reported by the float is actually in the ocean, or if by error it is reporting a position overland; test 18 checks whether the float is sending, again and again, the same data, etc. We don’t expect you to learn what every test is about, but if you are curious about all of them, take a look at the [Argo Quality Control Manual for CTD and Trajectory Data](https://archimer.ifremer.fr/doc/00228/33951/), and here there is a list of the RTQC tests:
 
@@ -40,7 +42,7 @@ In the following figure, and video,  we show the most common behaviours found in
 
 <center><img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/GIF_Bad_Data.gif" alt="Bad data" width="800"/></center>
 
-&nbsp;&nbsp;<center><iframe width="800" height="450"  src="https://drive.google.com/file/d/1SJXoCYij8muWN-zQVlD5pViSJKUJVjWJ/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>&nbsp;&nbsp;
[…]
