### 2024-08-15 euroargodev/argoonlineschool#13 @PedroVelez
> CXX refers to chapter number in Lesson 3 only.
>
> C10
>
> * Since it is an option to run the Jupyter Notebook locally, the instructions are insufficient to get started, which is OK because this is not the purpose of this lesson. However, it should be directly stated where they can start learning how. If it is the “An Introduction to Earth and Environmental Data Science” module, then you should state it.

Thanks a lot, we have pointed the reader to a module of “An Introduction to Earth and Environmental Data Science”

> C11
>
> * In [26] appears to hang for a very long time, and eventually did fail in the same way the other referee mentioned. If it is expected that some calculations will be slow, learners should be cautioned to rerun some lines (if run locally).
> * xrDS.close() has an error in the GitHub.io version.
> * Sst.set is used twice with little explanation of what the code will do (e.g., what is method=‘nearest’ do?)
> * In the section “Using the data”, you start this section by calculating he grid resolution but then don’t appear to use it.
> * In “Remote data access via openDAP”, not enough context is given. Too often lines of code are run without some complementary text.
>
> C21
>
> * The awesome image at the top should appear earlier in Lesson 1 and/or 2, and then again here.
> * This section is much more manageable as far as information loading than the “Using Argo Data” part.
> * “This is, within the cycle file, there are two profiles. The first one (N_PROF=0 in blue) is measured during its ascend from 2000 dbar to 5 dbar and it constitutes the core Argo program; the second one (N_PROF=1 in red) only measures the top 5 dbar.” Has dbar been explained at some point? Would most learners be familiar with it and know why it is used here instead of metres?
>
> C22
>
> * Colorbar label in TS plot is missing
>
> C23
>
> * Missing axis labels on bottom plot (unclear what is actually being plotted)
> * There should be more guiding text on this page explaining what you are doing here.
>
> C24
>
> * Last and second last figures don’t show very much. Is this intentional? Perhaps should be explained to a learner.
>
> C42
>
> * Missing some axis and colourbar labels in the first graph
> * Psal is all NaNs in output, I suspect this is not normal.
> * Second graph is empty

> CXX refers to chapter number in Lesson 3 only.
>
> C10
>
> * Since it is an option to run the Jupyter Notebook locally, the instructions are insufficient to get started, which is OK because this is not the purpose of this lesson. However, it should be directly stated where they can start learning how. If it is the “An Introduction to Earth and Environmental Data Science” module, then you should state it.
>
> C11
>
> * In [26] appears to hang for a very long time, and eventually did fail in the same way the other referee mentioned. If it is expected that some calculations will be slow, learners should be cautioned to rerun some lines (if run locally).
> * xrDS.close() has an error in the GitHub.io version.
> * Sst.set is used twice with little explanation of what the code will do (e.g., what is method=‘nearest’ do?)
> * In the section “Using the data”, you start this section by calculating he grid resolution but then don’t appear to use it.
> * In “Remote data access via openDAP”, not enough context is given. Too often lines of code are run without some complementary text.
>
> C21
>
> * The awesome image at the top should appear earlier in Lesson 1 and/or 2, and then again here.
> * This section is much more manageable as far as information loading than the “Using Argo Data” part.
> * “This is, within the cycle file, there are two profiles. The first one (N_PROF=0 in blue) is measured during its ascend from 2000 dbar to 5 dbar and it constitutes the core Argo program; the second one (N_PROF=1 in red) only measures the top 5 dbar.” Has dbar been explained at some point? Would most learners be familiar with it and know why it is used here instead of metres?
>
> C22
>
> * Colorbar label in TS plot is missing
>
> C23
>
> * Missing axis labels on bottom plot (unclear what is actually being plotted)
> * There should be more guiding text on this page explaining what you are doing here.
>
> C24
>
> * Last and second last figures don’t show very much. Is this intentional? Perhaps should be explained to a learner.
>
> C42
>
> * Missing some axis and colourbar labels in the first graph
> * Psal is all NaNs in output, I suspect this is not normal.
> * Second graph is empty
