### 2024-08-15 f717d3c
fixing JOSE Review: Lesson 3 - djmallum
