### 2024-08-16 b558071
Videos from L02 in YT
