+++ Lessons/L02_TheArgoData/Chapter25_GetAccesOrganized.md
@@ -4,21 +4,28 @@
 
 Let's see how the Argo data is organized in different folders in the GDACs:
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/1Ld73L9nTlebGpOdepRTDA3e5UwzPwQse/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+&nbsp;&nbsp;
+<center>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/jj0wnNDYOuY?si=PnMZLwdaWFVxrji3" title="how is the data organized?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+</center>
 
 
 #### What is FTP?
 
 But how we can actually get the data?:
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/1Mbfk07QANv4Jm7ZMTWbwNtbBxFcaemRM/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/akZeIJAc9e8?si=I05qI4Y11GCJVA32&amp;start=1" title="What is FTP?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+</center>
 
 
 #### Using FTP or HTTPs
 
 To access the data directly from the GDACs we can use the FTP or HTTPs protocols. Although it sound complex is very easy:
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/1YV1ZJajaPf0dh4KyvRx70dqTiWPdxVI0/preview" width="640" height="480" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/GNLrd6Efxk8?si=ssalirN3YdL-FDK7" title="Using FTP or HTTPs" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+</center>
 
 This are the address:
 
@@ -40,7 +47,8 @@ Take a look at this simple example on how to get access to Argo data through Cor
 3. You are now connected to the Coriolis GDAC!
 
 &nbsp;&nbsp;<center>
-<iframe width="800" height="450" src="https://drive.google.com/file/d/1Ct71C4wFOTvaL-J_oEUXa2-_rbYJmRKi/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
[…]
