+++ _sources/Lessons/L02_TheArgoData/Chapter24_QualityFlagMean.md
@@ -2,7 +2,9 @@
 
 Each data will have a specific quality flag assigned that indicates its status! Both Real-Time quality control and Delayed - Mode quality control follow a solid quality flag policy. Here we explain it, briefly:
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://drive.google.com/file/d/1iX3-8SGGAB95zyDQqxT9bVtH7ediBJJE/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>&nbsp;&nbsp;
+&nbsp;&nbsp;<center>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/KEfgHYIcCo4?si=FMwLghCknsWOSvvS&amp;start=2" title="What does a quality control flag mean?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+</center>&nbsp;&nbsp;
 
 And here there are some examples that show how robust this quality flag policy is:
 
