+++ _sources/Lessons/L01_TheArgoProgram/Chapter14_OceanographicProfile.md
@@ -1,15 +1,21 @@
 ## What is an oceanographic profile?
 
-Each of the temperature and salinity data sets measured from 2000 meters depth to surface is called an oceanographic profile, which is why the Argo floats are also known as "profilers".  Usually in oceanography observations are refered to the pressure level in decibar. The decibar is a multiple of the unit bar (bar) for pressure; the prefix deci (d) is a factor of one tenth or 0.1. Per this definition, one decibar is 0.1 bars. The decibar is represented by the symbol *dbar*. The bar is a metric unit of pressure defined as 100,000 Pa (100 kPa), though not part of the International System of Units (SI) is widely used in oceanography since you can get a rough idea of the depth in meters just looking the pressure in dbars and the rule of thumb than 1 meters is 1 dbar. 
+Each of the temperature and salinity data sets measured from 2000 meters depth to surface is called an oceanographic profile, which is why the Argo floats are also known as "profilers".
+
+Usually in oceanography, observations are refered to the pressure level in decibar. The decibar is a multiple of the unit bar (bar) for pressure; the prefix deci (d) is a factor of one tenth or 0.1. Therefore, one decibar is 0.1 bars. The decibar is represented by the symbol *dbar*. The dbar is a metric unit of pressure defined as 10 kPa, though not part of the International System of Units (SI) is widely used in oceanography since you can get a rough idea of the depth in meters just looking the pressure in dbars and the rule of thumb than 1 meters is 1 dbar.
 
 In the following video you will find a detailed explanation and some examples:
 
 
 &nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/m2hj1WrteCg?si=jPixZ_vXck0KAc3m&amp;start=6" title="What is an oceanographic profile?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
 
-Therefore and argo float will measure Conductivity, Temperature, and Pressure, this is, *CTD profile* and we refer to the sensor as a *CTD*. From the CTD we can obtain also the salinity.  Here there is an Sketch of a profile:
+Therefore and argo float will measure Conductivity, Temperature, and Pressure, this is, a *CTD profile*, and from the CTD measurements we can obtain also the salinity. Here there is an Sketch of a profile:
 
-<img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/ArgoCycleTH.png" alt="xarray logo" width="80"/>
+<center>
+<img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/ArgoCycleTH.png" alt="Argo cycle" width="800"/>
+</center>
+
+Usually in oceanography, the sensor that measures CTD it is also called a *CTD*.
 
 #### Argo has been revolutionary
 
