+++ intro.md
@@ -21,12 +21,12 @@ The AoS is organized in two parts. The fist one, organized in two lessons, is de
 
 In this section and using a set of videos you will learn about:
 * [What is the Argo network?](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter11_WhatArgoNetwork.html)
-* [Why do we need Argo and what are its objectives?](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter12_NeedArgo.html)
+* [Why Argo?](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter12_NeedArgo.html)
 * [How the observations are done?](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter13_ObservationsDone.html)
 * [What is an oceanographic profile?](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter14_OceanographicProfile.html)
 * [Technological innovations](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter15_TechInnovations.html)
-* [Recap Lesson 1](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter16_RecapLesson1.html)
-
+* [Recap](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter16_RecapLesson1.html)
+ 
 **2. [The Argo Data](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter10_TheArgoData_intro.html)**
 
 In this section and using a set of videos you will learn about:
@@ -35,7 +35,7 @@ In this section and using a set of videos you will learn about:
 * [QualityControl](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter14_QualityControl.html)
 * [What does a quality control flag mean?](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter15_QualityFlagMean.html)
 * [Accesing the Argo data](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter16_GetAccesOrganized.html) 
-* [Recap Lesson 2](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter17_RecapLesson2.html)
+* [Recap](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter17_RecapLesson2.html)
 
 Additonally we have included Quizzes for the [first](https://euroargodev.github.io/argoonlineschool/Lessons/Quizzes/Quiz10.html) and [second](https://euroargodev.github.io/argoonlineschool/Lessons/Quizzes/Quiz20.html) lessons if you want to challenge yourself! 
 
@@ -46,9 +46,10 @@ In this section and using hands-on jupyter notebooks you will learn about:
 * [Accesing Argo data by float](https://euroargodev.github.io/argoonlineschool/Lessons/L03_UsingArgoData/Chapter20_ArgoDatabyFloat_Intro.html)
 * [Accesing Argo data by date](https://euroargodev.github.io/argoonlineschool/Lessons/L03_UsingArgoData/Chapter30_ArgoDatabyDate_Intro.html)
 * [Aceesing Real-Time and Delayed mode data](https://euroargodev.github.io/argoonlineschool/Lessons/L03_UsingArgoData/Chapter40_RTandDM.html)
-* [Recap Lesson 3](https://euroargodev.github.io/argoonlineschool/Lessons/L03_UsingArgoData/Chapter43_RecapLesson3.html)
[…]
