+++ _sources/Lessons/L01_TheArgoProgram/Chapter16_RecapLesson1.md
@@ -1,4 +1,4 @@
-## Recap Lesson 1
+## Recap
 
 Let's summarize what you have learned from this lesson a about the Argo Data
 
