+++ _toc.yml
@@ -25,28 +25,28 @@ parts:
 
 - caption: Lesson 3 - Using Argo data
   chapters:
-  - file: Lessons/L03_UsingArgoData/Chapter10_UsingArgoData_intro.ipynb
+  - file: Lessons/L03_UsingArgoData/Chapter30_UsingArgoData_intro.ipynb
     sections:
-    - file: Lessons/L03_UsingArgoData/Chapter11_TheNetCDFFormat.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter31_TheNetCDFFormat.ipynb      
 
-  - file: Lessons/L03_UsingArgoData/Chapter20_ArgoDatabyFloat_Intro.ipynb      
+  - file: Lessons/L03_UsingArgoData/Chapter32_ArgoDatabyFloat_Intro.ipynb      
     sections:
-    - file: Lessons/L03_UsingArgoData/Chapter21_ArgoDatabyFloat_Cycle.ipynb      
-    - file: Lessons/L03_UsingArgoData/Chapter22_ArgoDatabyFloat_Prof.ipynb      
-    - file: Lessons/L03_UsingArgoData/Chapter23_ArgoDatabyFloat_Traj.ipynb      
-    - file: Lessons/L03_UsingArgoData/Chapter24_ArgoDatabyFloat_ArgoPy.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter32b_ArgoDatabyFloat_Cycle.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter32c_ArgoDatabyFloat_Prof.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter32d_ArgoDatabyFloat_Traj.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter32e_ArgoDatabyFloat_ArgoPy.ipynb      
 
-  - file: Lessons/L03_UsingArgoData/Chapter30_ArgoDatabyDate_Intro.ipynb      
+  - file: Lessons/L03_UsingArgoData/Chapter33_ArgoDatabyDate_Intro.ipynb      
     sections:
-    - file: Lessons/L03_UsingArgoData/Chapter31_ArgoDatabyDate_Day.ipynb      
-    - file: Lessons/L03_UsingArgoData/Chapter32_ArgoDatabyDate_ArgoPy.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter33b_ArgoDatabyDate_Day.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter33c_ArgoDatabyDate_ArgoPy.ipynb      
 
-  - file: Lessons/L03_UsingArgoData/Chapter40_RTandDM.ipynb      
+  - file: Lessons/L03_UsingArgoData/Chapter34_RTandDM.ipynb      
     sections:
-    - file: Lessons/L03_UsingArgoData/Chapter41_RTData.ipynb      
-    - file: Lessons/L03_UsingArgoData/Chapter42_DMData.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter34b_RTData.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter34c_DMData.ipynb      
 
-  - file: Lessons/L03_UsingArgoData/Chapter43_RecapLesson3.ipynb      
+  - file: Lessons/L03_UsingArgoData/Chapter35_RecapLesson3.ipynb      
     
 
 
