+++ _sources/Lessons/L03_UsingArgoData/Chapter30_UsingArgoData_intro.md
@@ -1,77 +1,0 @@
-# Using Argo Data
-
-As indicated in the introduction, the second part of the Argo Online School is hands-on, and it is mainly based on Jupyter Notebooks and a few short videos to teach, how to acces the data Argo data and use it.
-
-<center><iframe width="800" height="450" src="https://drive.google.com/file/d/17s5UwuNwyQzjL7EW9DYhWkAYP8nYTYLI/preview" title="The Argo Data" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></center>
-
-## The hands-on component of the Argo Online School
-
-You can access the hands-on content as here, in a web-page, built using [JupyterBook](https://euroargodev.github.io/argoonlineschool), or download the Jupyter Notebooks and use them in your local machine. 
-
-The hands-on component of the Argo Online School was developed using JupyterLab, since it contains a complete environment for interactive scientific computing which runs in your web browser. Jupyter is an open-source python project, and as a very useful first approach to python and JupyterLab, you can use [An Introduction to Earth and Environmental Data Science](https://earth-env-data-science.github.io/intro) and [Research computing in earth Sciences](https://rabernat.github.io/research_computing/) developed by [Ryan Abernathey](https://ocean-transport.github.io/) and [Kerry Key](https://emlab.ldeo.columbia.edu/index.php/team/kerry-key/).
-
-In case that you decide to download and run the python Jupyter Notebooks in your local machine, you should create a python environment. Since using some of the packages may give problems due to compatibility issues between conda-forge packages and packages contained in the default conda channels, we recommend to set up `channel_priority: strict` and give priority to the `conda-forge` channel over the default channels when installing stuff. There are two ways of doing it. Either you always specify `conda install -c conda-forge` or you create a .condarc file in your home with this content:
-```
-channels:
-  - conda-forge
-  - defaults
-channel_priority: strict
-```
-An environment should either use conda-forge or not, from creation to destruction. Do not mix and match. If you created it without using the conda-forge channel, then do not add it to the mix halfway. In practice, we always create an environment with conda-forge unless in very specific cases where we found incompatibilities.
-
-To create and activate the AOS environment, that already includes the last stable version of [argopy](https://argopy.readthedocs.io/en/latest/), the python library for Argo data, you should:
-
-```
-conda env create -n AOS -f environment.yml 
-conda activate AOS
-```
-
-If by any chance, a library is not included in the environment.yml file, you should install it using, for instance:
-
-```
-conda install netcdf4
-conda install xarray
-conda install seawater
-```
-
-For futher information about [Managing Python Environments](https://earth-env-data-science.github.io/lectures/environment/python_environments.html) we recommend the corresponding section in [An Introduction to Earth and Environmental Data Science](https://earth-env-data-science.github.io/intro)
-
-
-### Data used in the Argo Online School
-
-If you download the notebooks you should create a ./Data* folder to include the data files used for the examples. The files in the Data folder can be downloaded directly from its source using _wget_, for instance:
-
-for the *Daily NOAA OI SST V2 High-Resolution Dataset for 2020*:
-
-```
-wget ftp://ftp2.psl.noaa.gov/Datasets/noaa.oisst.v2.highres/sst.day.mean.2020.nc
-```
-
-for the data from float 6901254:
-
-```
-wget ftp://ftp.ifremer.fr/ifremer/argo/dac/coriolis/6901254/*`
-```
-
-or for the data in the North Atlantic for the 11th November 2019
-
[…]
