+++ Lessons/L02_TheArgoData/Chapter20_TheArgoData_intro.md
@@ -7,7 +7,9 @@
 
 In this lesson, you will get a deep approach to Argo data: Its basic features, where it comes from, and how it is organized for proper and successful management. It is a big dataset! 
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/4oM2ZY46ETA?si=B2jZYtS7Wb6iBWeL" title="A big data set" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>&nbsp;&nbsp;
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/4oM2ZY46ETA?si=B2jZYtS7Wb6iBWeL" title="The Argo Online School 201 - The Argo data.  Introduction: A big dataset" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+_The Argo Online School 201 - The Argo data. Introduction: A big dataset_
+&nbsp;&nbsp;
 
 ### The WMO float identifier: a unique number for each float
 
