+++ _sources/Lessons/L01_TheArgoProgram/Chapter14_OceanographicProfile.md
@@ -7,7 +7,10 @@ Usually in oceanography, observations are refered to the pressure level in decib
 In the following video you will find a detailed explanation and some examples:
 
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/m2hj1WrteCg?si=jPixZ_vXck0KAc3m&amp;start=6" title="What is an oceanographic profile?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+&nbsp;&nbsp;
+<center><iframe width="800" height="450" src="https://www.youtube.com/embed/m2hj1WrteCg?si=jPixZ_vXck0KAc3m&amp;start=6" title="What is an oceanographic profile?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+_The Argo Online School 141 - The Argo Program. What is an oceanographic profile?._
+&nbsp;&nbsp;
 
 Therefore and argo float will measure Conductivity, Temperature, and Pressure, this is, a *CTD profile*, and from the CTD measurements we can obtain also the salinity. Here there is an Sketch of a profile:
 
@@ -21,4 +24,9 @@ Usually in oceanography, the sensor that measures CTD it is also called a *CTD*.
 
 Argo represents a revolution in ocean observation, since it provides data, freely and free of charge, at a  rate 20 times higher compared to traditional methods on board oceanographic records. Here we present some more interesting facts:
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/HjOG0TgMVGo?si=UAmTlAanhANDvCG3&amp;start=3" title="Argo has been revolutionary" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+&nbsp;&nbsp;
+<center>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/HjOG0TgMVGo?si=UAmTlAanhANDvCG3&amp;start=3" title="The Argo Online School 142 - The Argo Program.  What is an oceanographic profile?: Argo has been revolutionary" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+</center>
+_The Argo Online School 142 - The Argo Program. What is an oceanographic profile?: Argo has been revolutionary_
+&nbsp;&nbsp;
