+++ _sources/Lessons/L02_TheArgoData/Chapter25_GetAccesOrganized.md
@@ -1,44 +1,50 @@
-# Accesing the Argo data
+## Accesing the Argo data
 
-#### how is the data organized?
+### how is the data organized?
 
 Let's see how the Argo data is organized in different folders in the GDACs:
 
 &nbsp;&nbsp;
 <center>
-<iframe width="800" height="450" src="https://www.youtube.com/embed/jj0wnNDYOuY?si=PnMZLwdaWFVxrji3" title="how is the data organized?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/jj0wnNDYOuY?si=PnMZLwdaWFVxrji3" title="The Argo Online School 251 - The Argo data.  Accesing the Argo data: What is FTP?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 </center>
+_The Argo Online School 251 - The Argo data. Accesing the Argo data: What is FTP?_
+&nbsp;&nbsp;
 
-
-#### What is FTP?
+### What is FTP?
 
 But how we can actually get the data?:
 
-&nbsp;&nbsp;<center>
-<iframe width="800" height="450" src="https://www.youtube.com/embed/akZeIJAc9e8?si=I05qI4Y11GCJVA32&amp;start=1" title="What is FTP?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+&nbsp;&nbsp;
+<center>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/akZeIJAc9e8?si=I05qI4Y11GCJVA32&amp;start=1" title="The Argo Online School 252 - The Argo data.  Accesing the Argo data: What is FTP?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 </center>
+_The Argo Online School 252 - The Argo data. Accesing the Argo data: What is FTP?_
+&nbsp;&nbsp;
 
-
-#### Using FTP or HTTPs
+### Using FTP or HTTPs
 
 To access the data directly from the GDACs we can use the FTP or HTTPs protocols. Although it sound complex is very easy:
 
-&nbsp;&nbsp;<center>
-<iframe width="800" height="450" src="https://www.youtube.com/embed/GNLrd6Efxk8?si=ssalirN3YdL-FDK7" title="Using FTP or HTTPs" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+&nbsp;&nbsp;
+<center>
+<iframe width="800" height="450" src="https://www.youtube.com/embed/GNLrd6Efxk8?si=ssalirN3YdL-FDK7" title="The Argo Online School 253 - The Argo data.  Accesing the Argo data: Using FTP or HTTPs" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
[…]
