### 2024-08-26 #193 @PedroVelez
Hi, thanks a lot for the patience @kls2177 we have all most ended including all the suggestions , we have replied to each of of them in the issues opened for the repository:

[Typos](https://github.com/euroargodev/argoonlineschool/issues/10)
[Lesson 1](https://github.com/euroargodev/argoonlineschool/issues/11)
[Lesson 2](https://github.com/euroargodev/argoonlineschool/issues/12)
[Lesson 3](https://github.com/euroargodev/argoonlineschool/issues/13)
[Checklist Comments (See above)](https://github.com/euroargodev/argoonlineschool/issues/14)
[General Comments](https://github.com/euroargodev/argoonlineschool/issues/15)

and we would like to know what do we have to do for re-submission.

thanks
pedro
