+++ paper.bib
@@ -1,17 +1,19 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
-
 %% Created for Alberto Gonzalez at 2023-07-24 16:20:50 +0100 
-
-
 %% Saved with string encoding Unicode (UTF-8) 
 
 
-
-@incollection{cite-key,
-	crossref = {ArgoQCManual2022},
-	date-added = {2023-07-24 16:20:08 +0100},
-	date-modified = {2023-07-24 16:20:08 +0100}}
+@techreport{ArgoManual2022,
+	title = "Argo user’s manual",
+	type = "Report",
+	url = "https://doi.org/10.13155/29825",
+	doi = "10.13155/29825",
+	author = "Argo data management",
+	year = "2022",
+	editor = "Ifremer",
+	abstract = "This document is the Argo data user’s manual. It contains the description of the formats and files produced by the Argo Data Assembly Centres (DACs)."
+}
 
 @misc{Abernathey2021,
 	author = {Ryan Abernathey, Kerry Key and Tim Crone},
@@ -98,16 +100,6 @@
 	url = {https://doi.org/10.13155/33951},
 	year = {2022},
 	bdsk-url-1 = {https://doi.org/10.13155/33951}}
-
-@techreport{ArgoUserManual2021,
-	author = {Argo Data Management Team},
-	institution = {Argo},
-	month = {01},
-	organization = {Argo Data Management Team},
-	title = {Argo users manual},
-	url = {https://doi.org/10.13155/29825},
-	year = {2021},
-	bdsk-url-1 = {https://doi.org/10.13155/29825}}
 
 @article{Argovis,
 	address = {Boston MA, USA},
