+++ paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-The **Argo Online School (AoS)** is a set of videos, animations and hands-on python driven jupyter notebooks designed to make the data, of the more than 4000 profiling floats that constitute the Argo program, accessible. The Argo program samples, in near real time, the upper 2000 metres of the ocean using a fleet of floats that drift with the ocean currents. The AoS is integrated by 25 sections about the Argo program and its data, organized into a total of 3 lessons and 2 self-assessment sections:
+The **Argo Online School (AoS)** is a set of videos, animations and hands-on python driven jupyter notebooks designed to make the data from the more than 4000 profiling floats that constitute the Argo program, accessible. The Argo program samples, in near real time, the upper 2000 metres of the ocean using a fleet of floats that drift with the ocean currents. The AoS is integrated by 28 chapters about the Argo program and its data, organized into a total of 3 lessons and 2 self-assessment sections:
 
 * Lesson 1 is intended to introduce the basics of Argo, its objectives and key elements, such as the structure of the Argo floats and their operation in the open ocean.
 
@@ -30,9 +30,7 @@ The **Argo Online School (AoS)** is a set of videos, animations and hands-on pyt
 
 * Lesson 3 is the AoS hands-on component, and it requieres some knowledge of python. It offers a set of instructions to prepare a python environment in case the user wants to run the python Jupyter Notebooks in a local machine. This environment already includes recommended packages. The walkthrough of Lesson 3 shows how to work with the netCDF format, how to access and process Argo data, and how to produce fancy plots for understanding information from Argo data.
 
-Lessons 1 and 2 are itself a closed module and could be used by a learner without using lesson 3, assuming the goal is simply to learn about Argo.
-
-The target audience of the AoS are high school, undergraduate or early graduate students, and the programming content included in Lesson 3 provides an ideal opportunity to support students with technical or science curriculum. It is recommended not to skip any lesson in the AoS as the content is carefully connected from less to greater complexity, guaranteeing a progressive learning. The user has the opportunity to self-assess the learning process carried out during their course in the AoS through the proposed quizzes.
+The target audience of the AoS are high school, undergraduate or early graduate students, and the programming content included in Lesson 3 provides an ideal opportunity to support students with technical or science curriculum. Lessons 1 and 2 are itself a closed module and could be used by a learner without using lesson 3, assuming the goal is simply to learn about Argo. It is recommended not to skip any lesson in the AoS as the content is carefully connected from less to greater complexity, guaranteeing a progressive learning. The user has the opportunity to self-assess the learning process carried out during their course in the AoS through the proposed quizzes. 
 
 The AoS has been designed to have the possibility to be expanded to follow the implementation of new features in the Argo program.
 
@@ -48,24 +46,26 @@ The AoS was presented at the [2nd Ocean Observers Workshop](https://bit.ly/3pUCh
 
 # Notes on instructional design
 
-The AoS aims to teach the basic foundations to understand and use the Argo data. The AoS does not attempt to show everything about the Argo program, since the Argo documentation is available from the Argo Data Managment Team ([http://www.argodatamgt.org/Documentation](http://www.argodatamgt.org/Documentation)) [@ArgoQCManual2022; @ArgoUserManual2021] for deeper learning. The AoS does not pretend to be a library or Application Programming Interface to ease Argo data access, since there are other initiatives with that purpose; among others, Argopy, a python library for Argo data beginners and experts [@Maze2020] or  ArgoVis, a web application for fast delivery, visualization, and analysis of argo Data [@Argovis].
[…]
