+++ _sources/intro.md
@@ -6,18 +6,19 @@
 
 # The Argo Online School
 
-Argo is an international program that collects information from inside the ocean using a fleet of floats that drift with the ocean currents. The floats move up and down between the surface and a mid-water level measuring ocean variables but spend almost all their life below the surface. Argo has been collecting more than 100,000 profiles per year since 2012, and nowadays [Argo](https://argo.ucsd.edu/) is the major component of the Global Ocean Observing System.
+[Argo](https://argo.ucsd.edu/) is an international program that collects information from inside the ocean using a fleet of floats that drift with the ocean currents. The floats move up and down between the surface and a mid-water level, measuring ocean variables, but spend almost all their life below the surface. Argo has been collecting more than 100,000 profiles per year since 2012, and nowadays Argo is the major component of the [Global Ocean Observing System](https://goosocean.org/).
 
 <img src="http://sio-argo.ucsd.edu/statusbig.gif" alt="Argo status" width="800"/>
+&nbsp;&nbsp;
 
-Given the amount of data gathered by the Argo network of floats, the associated documentation has grown considerably and can be overwhelming. 
-
-That's why we developed the Argo Online school ([AoS](https://euroargodev.github.io/argoonlineschool/)), to teach the basic foundations to use and understand the Argo data. The AoS does not attempt to show everything, since the Argo documentation is available for deeper learning. The AoS is a set of short videos and hands-on jupyter notebooks designed to be accessible for high school or graduate students in any discipline, with no prerequisites.
-
-The AoS is organized in two parts. The fist one, organized in two lessons, is devoted to teach what is the [The Argo Program](https://euroargodev.github.io/argoonlineschool/Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.html) and the [The Argo Data](https://euroargodev.github.io/argoonlineschool/Lessons/L02_TheArgoData/Chapter10_TheArgoData_intro.html) , with a set of short videos. The first part, this is Lessons 1-2 are itself a closed module and could be used by a learner without using lesson 3, assuming the goal is simply to learn about Argo. The second part includes only one lesson: [Using the Argo data](https://euroargodev.github.io/argoonlineschool/Lessons/L03_UsingArgoData/Chapter10_UsingArgoData_intro.html)**, with jupyter notebooks and short videos to teach, hands-on, how to acces the data Argo data and use it.
+Given the amount of data gathered by the Argo network of floats, the associated documentation has grown considerably and can be overwhelming. That's why we developed the Argo Online school (AoS), to teach the basic foundations to use and understand the Argo data. The AoS is a set of guided-short videos and hands-on jupyter notebooks designed to for high school or graduate students in any discipline.
  
+The AoS is organized in two parts. The fist one, wth two lessons, is devoted to teach what is the The Argo Program and The Argo Dataset, with a set of 32 short videos and 14 chapters. These two lessons are itself a closed module, assuming the goal is simply to learn about Argo. The second part includes only one lesson: Using the Argo data, with jupyter notebooks and 8 short videos to teach, hands-on in 14 chapters, how to acces the data Argo data and use it.  Lessons 1 and 2 require about 5 hours to be completed, while lesson 3 requires 10 hours to be completed, assuming basic knowledge of python.
 
[…]
