### 2024-08-28 euroargodev/argoonlineschool#14 @PedroVelez
Hi @djmallum, thanks! we have fixed all the suggestions listed in here.
