### 2024-08-28 euroargodev/argoonlineschool#15 @PedroVelez
Hi @djmallum, thanks! we have fixed all the typos/suggestions listed in here.
