+++ paper.md
@@ -14,7 +14,7 @@ authors:
    orcid: 0000-0003-2404-5679
    affiliation: 1
 affiliations:
-- name: Centro Oceanográfico de Canaarias, Instituto Español de Oceanografía (IEO - CSIC)
+- name: Centro Oceanográfico de Canarias, Instituto Español de Oceanografía (IEO - CSIC)
   index: 1
 date: 1 April 2022
 bibliography: paper.bib
@@ -40,7 +40,7 @@ Argo [@roemmich-2019; @riser-2016] is an international program that collects inf
 
 Although data from Argo [@Argo2021] is publicly and freely available, it is a huge dataset, with thousands of files and the free documentation is rather complex since it has to cope with the many different models of floats and sensors, and the traceability of the quality control of all the measurements. The Argo community has always been aware of the difficulties of early users to manage the complex and large datasets from the network. In that sense, the Argo Online School addresses an existing need for the Argo community to communicate on the Argo programme and how its invaluable dataset can be used for various applications.
 
-The potential of e-learning platforms is used in the Argo Online School to offer all kinds of resources to users and thus promote and improve the access and use of the Argo data. In this way, the AoS is defined as an e-learning tool based on an interactive environment similar to other already established educational platforms. It is a tool with great educational potential that not only seeks to show the basic steps to use Argo data, but also to empower users so that they face future _programming_ scenarios in their academic and professional stage. 
+The potential of e-learning platforms is used in the Argo Online School to offer all kinds of resources to users and thus promote and improve the access and use of the Argo data. In this way, the AoS is defined as an e-learning tool based on an interactive environment similar to other already established educational platforms. It is a tool with great educational potential that not only seeks to show the basic steps to use Argo data, but also to empower users so that they face future _programming_ scenarios in their academic and professional stage.
 
 The AoS was presented at the [2nd Ocean Observers Workshop](https://bit.ly/3pUChmJ) on November 20th, 2021  and at the [22nd Argo Data Management Team meeting](https://bit.ly/3e39rLL) on December 10, 2021.
 
@@ -59,11 +59,11 @@ The AoS is a set of videos and hands-on python driven jupyter notebooks, designe
 
 All content is divided into three lessons: 1. The Argo Program, that describes the basic concepts of this ocean observing network; 2. The Argo Data, that describes how the data is organized; and 3. Using the Argo data, that uses the knowledge of the previous sections and python driven jupyter notebooks to teach how to use the data. Finally, a quiz section is included for self-appraisal. Lessons 1 and 2 are aimed at users with minimal or no knowledge of the Argo network, therefore no prerequisites are needed, and itself are a closed module and could be used by a learner without using lesson 3, assuming the goal is simply to learn about Argo. Lessons 1 and 2 contain 32 short videos and 14 chapters and require about 5 hours to be completed. Lesson 3 is intended for advanced users, as it requires basic programming skills in python. However, lesson 3 is duly explained step by step, to facilitate the transition of users coming from lessons 1 and 2. The basic recommendations and instructions for configuring the hands-on section of the AoS are also provided, whether the users want to work online or if they want to work on their local computer. Specific libraries and packages are recommended to guarantee the correct functioning of the AoS. Lessons 3 contains 8 short videos and 14 chapters and require about 10 hours to be completed, assuming basic knowledge of python.
 
-AQUI incluir como se puede adaptar a su suso.  potential uses.Does the documentation explain how someone would adopt the module, and include examples of how to use it?_
-
[…]
