+++ _sources/Lessons/L01_TheArgoProgram/Chapter10_TheArgoProgram_intro.md
@@ -30,5 +30,13 @@ _The Argo Online School 102 - The Argo Program. Introduction: Why study the ocea
 
 Marine environment data is a valuable asset. Quick access to reliable and accurate information is vital to address threats to the marine environment, in the development of policies and laws to protect vulnerable areas of our coasts and oceans, to understand trends, and to forecast future changes. Unfortunately, the collection, storage and access to marine data has been fragmented for many years.
 
-&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/7P2Rqw52PXk?si=_awesB9L8Sh1S61y" title="The Argo Online School 103 - The Argo Program.  Introduction: A new way to observe the ocean" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
+&nbsp;&nbsp;<center><iframe width="800" height="450" src="https://www.youtube.com/embed/7P2Rqw52PXk?si=_awesB9L8Sh1S61y" title="The Argo Online School 103 - The Argo Program.  Introduction: A new way to observe the ocean" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>&nbsp;&nbsp;
+
+###  Why is called Argo
+
+In Greek mythology, Argo was the ship in which Jason and the Argonauts set sail to search for the golden fleece. Argo floats sail the 21st century seas and Argo complements the satellite mission called *[Jason](https://www.eumetsat.int/our-satellites/jason-series)*, that provide global sea surface height observations for climate monitoring and ocean forecasts. Data from Argo and Jason complemente each other.  At the very begging of developing Argo, it was called the **AR**ray of **G**eostrophic **O**ceanography, since the profiling floats measuring temperature and salinity, at depth, together with the Jason satellites, could provide, by using the mathematical aproach called geostrophy, ocean currents in all the oceans.
+
+
+
+
 _The Argo Online School 103 - The Argo Program. Introduction: A new way to observe the ocean_
