### 2024-08-28 5032fcf
update bibliography
