### 2024-08-28 #193 @PedroVelez
> @Alberto-GS and @PedroVelez, the reviewers have completed their reviews and have comments for your consideration. I recommend that these comments be addressed before publication.
>
> Thanks @francispoulin and @djmallum for your helpful comments. @djmallum, quick question - I do not see an issue on the argoonlineschool repo entitled "General Comments". Not sure if this got included elsewhere.
>
> @francispoulin and @djmallum, would you be willing to review the revised version when available?
>
> Just a few of my own comments: I agree with @francispoulin, that it would be nice to integrate some python examples earlier in the course. Perhaps, just some simple plotting examples - rather that providing some static images, ask the students to generate the plots using python. This would allow students who are learning python for the first time to get some gradual exposure before Lesson 3.

- [x] Thanks for the comment, but the way the AoS has been designed the lessons 1 and 2 are a module by itself, that do not have any requirements for high school students , and we want to keep in that way. using some python code, for instance to plot the active floats shown in the first lesson may scare student without programming background. We have eliminated the 'with no prerequisites' statements
- [x]  on the other hand we have modified the Lesson 3, and we think is more gradual, and we point the students to basic python courses.

>
> You mention that this material is for high school or graduate students - what about undergraduate students?

- [x] Yes, it is also for undergrads, we now mention it on the paper and the AoS.
>
> I do not necessarily agree with your statement that no pre-requisite knowledge is required. There are certainly expectations about general science and mathematics knowledge. I suggest providing some information about pre-requisite knowledge on the landing page of the Argo Online School website. For example, what pre-requisite oceanography, maths, chemistry, geography knowledge is expected of students?

- [X] Thanks, we have removed the statement about  no pre-requisit, in the paper and the landing page, and modified the text to show what it is necessary, basically the knowledge obtained in middle school.
-
>This is likely more relevant for high school students than post-secondary students. At the beginning of Lesson 3, you provide students with a link to the Earth and Environmental Data Science course to learn python if they need to, but it would be nice to provide students with information about these pre-requisite coding skills and resources at the beginning of the course.

- [X] Thanks, we warn the students in the landing page and also in the JOSE paper, that lesson 3 requieres basic knownledge of python. We do not ask for xarray or numpy, since once you understand  python, either from the AoS or the Earth and Environmental Data Science course yoy can learn the basics of xarray and numpy.

>
> I really like the interactive quizzes! These look great and give timely feedback to students. I would recommend integrating a few more of these at regular intervals to keep the students engaged. Structurally, I would also suggest embedding the quizzes into each lesson, rather than having them as a separate section at the end. I feel that this would better reflect the sequencing of the material.
>
- [X] Thanks!, we have moved each quizz to the end of each lesson.
