+++ environment.yml
@@ -3,214 +3,214 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - anyio=4.4.0=pyhd8ed1ab_0
-  - appnope=0.1.4=pyhd8ed1ab_0
-  - argon2-cffi=23.1.0=pyhd8ed1ab_0
-  - argon2-cffi-bindings=21.2.0=py312h02f2b3b_4
-  - arrow=1.3.0=pyhd8ed1ab_0
-  - asttokens=2.4.1=pyhd8ed1ab_0
-  - async-lru=2.0.4=pyhd8ed1ab_0
-  - attrs=24.2.0=pyh71513ae_0
-  - babel=2.14.0=pyhd8ed1ab_0
-  - beautifulsoup4=4.12.3=pyha770c72_0
-  - bleach=6.1.0=pyhd8ed1ab_0
-  - blosc=1.21.6=h5499902_0
-  - brotli=1.1.0=hb547adb_1
-  - brotli-bin=1.1.0=hb547adb_1
-  - brotli-python=1.1.0=py312h9f69965_1
-  - bzip2=1.0.8=h99b78c6_7
-  - c-ares=1.33.0=h99b78c6_0
-  - ca-certificates=2024.7.4=hf0a4a13_0
-  - cached-property=1.5.2=hd8ed1ab_1
-  - cached_property=1.5.2=pyha770c72_1
-  - cartopy=0.23.0=py312h8ae5369_1
-  - certifi=2024.7.4=pyhd8ed1ab_0
-  - cffi=1.17.0=py312h80c9ed6_0
-  - cftime=1.6.4=py312hbebd99a_0
-  - charset-normalizer=3.3.2=pyhd8ed1ab_0
-  - comm=0.2.2=pyhd8ed1ab_0
-  - contourpy=1.2.1=py312h0fef576_0
-  - cycler=0.12.1=pyhd8ed1ab_0
-  - debugpy=1.8.5=py312h5c2e7bc_0
-  - decorator=5.1.1=pyhd8ed1ab_0
-  - defusedxml=0.7.1=pyhd8ed1ab_0
-  - entrypoints=0.4=pyhd8ed1ab_0
-  - exceptiongroup=1.2.2=pyhd8ed1ab_0
-  - executing=2.0.1=pyhd8ed1ab_0
-  - fonttools=4.53.1=py312h7e5086c_0
-  - fqdn=1.5.1=pyhd8ed1ab_0
-  - freetype=2.12.1=hadb7bae_2
-  - geos=3.12.2=h00cdb27_1
-  - gsw=3.6.19=py312hbebd99a_0
-  - h11=0.14.0=pyhd8ed1ab_0
-  - h2=4.1.0=pyhd8ed1ab_0
-  - hdf4=4.2.15=h2ee6834_7
-  - hdf5=1.14.3=nompi_hec07895_105
-  - hpack=4.0.0=pyh9f0ad1d_0
-  - httpcore=1.0.5=pyhd8ed1ab_0
-  - httpx=0.27.0=pyhd8ed1ab_0
-  - hyperframe=6.0.1=pyhd8ed1ab_0
-  - icu=75.1=hfee45f7_0
-  - idna=3.7=pyhd8ed1ab_0
-  - importlib-metadata=8.2.0=pyha770c72_0
-  - importlib_metadata=8.2.0=hd8ed1ab_0
-  - importlib_resources=6.4.2=pyhd8ed1ab_0
-  - ipykernel=6.29.5=pyh57ce528_0
-  - ipython=8.26.0=pyh707e725_0
-  - isoduration=20.11.0=pyhd8ed1ab_0
[…]
