### 2024-09-02 #193 @PedroVelez
Hi @francispoulin @kls2177 @djmallum we haver finished to do the review, including all the suggestions, typos and errors. Please could you let us now what is the next step?
best
pedro
