### 2024-09-06 #193 @francispoulin
I did a git pull of the repo and see that there is a problem in Chapter32.

When I execute the second cell I get a KeyError.

Can the authors please check if this works for them in the freshly pulled repo?
