### 2024-09-06 ac8230e
Modified the data structure to Data/202107-ArgoData/
We have done theses changes to reproduce how the data would be downloaded from the Argo snapshoots
