### 2024-09-06 afea6e5
added examples
