+++ _toc.yml
@@ -51,7 +51,7 @@ parts:
   - file: Lessons/L03_UsingArgoData/Chapter36_Examples.ipynb      
     sections:
     - file: Lessons/L03_UsingArgoData/Chapter36a_Status.ipynb      
-    - file: Lessons/L03_UsingArgoData/Chapter36b_TrayectoriasArea.ipynb      
+    - file: Lessons/L03_UsingArgoData/Chapter36b_TrayectoriesArea.ipynb      
 
 
 
