### 2024-09-06 #193 @PedroVelez
Thanks for reviewing it!

There was a problem with the numpy version and xarray, therefore the numpy version has to be 1.26.4 until xarray fix the problem.

We have updated the [environment.yml](https://github.com/euroargodev/argoonlineschool/blob/master/environment.yml) and verified that a clean installation of the enviroment, using
```
conda env create -n AOS -f environment.yml
conda activate AOS

```
works.  I know is very inconvenient, but in my experience this is often the case of python.

By the way, following some suggestion we have just added a section with more  [examples](https://euroargodev.github.io/argoonlineschool/Lessons/L03_UsingArgoData/Chapter36_Examples.html)  including one very similar to the [ArgoStatus](http://sio-argo.ucsd.edu/statusbig.gif)

best
pedro

> I did a git pull of the repo and see that there is a problem in Chapter32.
>
> When I execute the second cell I get a KeyError.
>
> Can the authors please check if this works for them in the freshly pulled repo?
