+++ Lessons/L01_TheArgoProgram/Chapter16_RecapLesson1.md
@@ -6,6 +6,7 @@ Let's summarize what you have learned from this lesson a about the Argo Data
 <center>
     <iframe width="800" height="450" src="https://www.youtube.com/embed/pqiC8i13KFY?si=hDej9MABw-eJsxmW&amp;start=9" title="Recap Lesson 1" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 </center>
+
 _The Argo Online School 161 - The Argo Program. Recap Lesson 1_
 &nbsp;&nbsp;
 
