+++ Lessons/L01_TheArgoProgram/Chapter14_OceanographicProfile.md
@@ -9,13 +9,17 @@ In the following video you will find a detailed explanation and some examples:
 
 &nbsp;&nbsp;
 <center><iframe width="800" height="450" src="https://www.youtube.com/embed/m2hj1WrteCg?si=jPixZ_vXck0KAc3m&amp;start=6" title="What is an oceanographic profile?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe></center>
-_The Argo Online School 141 - The Argo Program. What is an oceanographic profile?._
+
+_The Argo Online School 141 - The Argo Program. What is an oceanographic profile?_
 &nbsp;&nbsp;
 
 Therefore and argo float will measure Conductivity, Temperature, and Pressure, this is, a *CTD profile*, and from the CTD measurements we can obtain also the salinity. Here there is an Sketch of a profile:
 
 <center>
+<figure>
 <img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/ArgoCycleTH.png" alt="Argo cycle" width="800"/>
+<figcaption>Behaviour of an Argo float</figcaption>
+</figure>
 </center>
 
 Usually in oceanography, the sensor that measures CTD it is also called a *CTD*.
@@ -28,5 +32,6 @@ Argo represents a revolution in ocean observation, since it provides data, freel
 <center>
 <iframe width="800" height="450" src="https://www.youtube.com/embed/HjOG0TgMVGo?si=UAmTlAanhANDvCG3&amp;start=3" title="The Argo Online School 142 - The Argo Program.  What is an oceanographic profile?: Argo has been revolutionary" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 </center>
+
 _The Argo Online School 142 - The Argo Program. What is an oceanographic profile?: Argo has been revolutionary_
 &nbsp;&nbsp;
