+++ _sources/Lessons/L02_TheArgoData/Chapter21_ArgoDataStream.md
@@ -30,5 +30,6 @@ The “final data pool” is complemented with the inclusion of a historical cli
 &nbsp;&nbsp;<center>
 <iframe width="800" height="450" src="https://www.youtube.com/embed/hDhV80TB_QA?si=yEkEcxpBKwW-2KwA&amp;start=1" title="The Argo Online School 211 - The Argo data.  The Argo Data flow: All the data is saved" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 </center>
+
 _The Argo Online School 211 - The Argo data. The Argo Data flow: All the data is saved_
 &nbsp;&nbsp;
