+++ _sources/Lessons/L02_TheArgoData/Chapter24_QualityFlagMean.md
@@ -6,6 +6,7 @@ Each data will have a specific quality flag assigned that indicates its status! 
 <center>
 <iframe width="800" height="450" src="https://www.youtube.com/embed/KEfgHYIcCo4?si=FMwLghCknsWOSvvS&amp;start=2" title="The Argo Online School 241 - The Argo data.  What does a quality control flag mean?: how is the data organized?" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 </center>
+
 _The Argo Online School 241 - The Argo data. What does a quality control flag mean?: how is the data organized?_
 &nbsp;&nbsp;
 
