+++ _sources/Lessons/L02_TheArgoData/Chapter26_RecapLesson2.md
@@ -6,5 +6,6 @@ Let's summarize what you have learned from this lesson a about the Argo Data
 <center>
 <iframe width="800" height="450" src="https://www.youtube.com/embed/K5Kh2ztOuUc?si=z8VAtfCrsiJ96fF4&amp;start=1" title="The Argo Online School 261 - The Argo data. Recap Lesson 2." frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 </center>
+
 _The Argo Online School 261 - The Argo data. Recap Lesson 2._
 &nbsp;&nbsp;
