+++ _sources/Lessons/L01_TheArgoProgram/Chapter15_TechInnovations.md
@@ -6,6 +6,7 @@ Argo is constantly trying to improve the quality and coverage of its data, which
 &nbsp;&nbsp;<center>
 <iframe width="800" height="450" src="https://www.youtube.com/embed/UVUkdOgFafU?si=XGC59V7C9Yhbao-M" title="The Argo Online School 151 - The Argo Program.  Technological innovations." frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 </center>
+
 _The Argo Online School 151 - The Argo Program. Technological innovations_
 &nbsp;&nbsp;
 
