### 2024-09-06 #193 @PedroVelez
> Thanks for looking into this.
>
> I don't use conda (and don't want to start).
>
> I checked on my ubuntu machine and I am using numpy 1.22.4. I did a git pull and I still had the same error, probably for the reasons you suggested.
>
> I don't know if I can proceed with my review is the library does not run on my machine.
>
> @kls2177 , any suggestions on how best to proceed?

How do you manage environments in python  Any method would allow you to update to the need version
