### 2024-09-06 #193 @francispoulin
I also noticed in Chapter 31 I got a RuntimeError in `dataOPsubsampled.sst.values()`.
