### 2024-09-06 #193 @francispoulin
> > Thanks for looking into this.
> > I don't use conda (and don't want to start).
> > I checked on my ubuntu machine and I am using numpy 1.22.4. I did a git pull and I still had the same error, probably for the reasons you suggested.
> > I don't know if I can proceed with my review is the library does not run on my machine.
> > @kls2177 , any suggestions on how best to proceed?
>
> How do you manage environments in python Any method would allow you to update to the need version

Thanks @PedroVelez .

I have not used environment variables in python for years so I am very rusty.  Please bare with me here.

I have learned that I have PyYAML installed and this should allow me to `import yaml` and then load the envrionment.  Does that sound right to you?

I did a simple test and didn't get an error

```
>>> import yaml
>>> open('environment.yml')
<_io.TextIOWrapper name='environment.yml' mode='r' encoding='UTF-8'>
```

however when I tried excuting your line with `prof = ...` I still got an error.

Any advice on what I should try?
