### 2024-09-06 #193 @PedroVelez
> I also noticed in Chapter 31 I got a RuntimeError in `dataOPsubsampled.sst.values()`.

HI, thanks.

We advice the readers that "However, be aware that sometimes remote access could be very slowly or even fail", and lately the NOAA server that provides with the OPEnDAP - thredds data is very unstable.

We have modifed the code, and now it seems more stable, however, If this is a problem we could eliminate the section "Remote data access via openDAP". Honestly I think it does not add anything and it give a lot of problems.
