### 2024-09-13 #193 @kls2177
Hi @PedroVelez and @Alberto-GS,

   If @djmallum is able to successfully run the code using the specified conda environment, then I am comfortable  proceeding with the assumption that most users will be using conda. We will wait for his comments and go from there. Thanks for your patience.
