### 2024-09-13 #193 @djmallum
Hi, I have been able to run the notebook with conda without issue. I was not able to dedicate enough time to compile a complete list of comments this week. I am out of office until Friday next week and so it is unlikely that I will have comments before the week of the 23rd.
