### 2024-10-04 #193 @Alberto-GS
> Great! Thank you so much @francispoulin and @djmallum for your reviews.
>
> @Alberto-GS, let's get this published! Please create an archive (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other) and post the archive DOI this REVIEW issue. Thanks!

Thanks a lot! These are great news! :)) We'll take a look to this archive creation on Zenodo. If we have any issues, we'll let you know guys. Just a silly question, since it's our first time: Does the repository will share the same DOI (JOSE and Zenodo)? I guess a DOI will be generated when the repo is archived in Zenodo?
