### 2024-10-04 #193 @Alberto-GS
Here is the archive DOI https://zenodo.org/records/13628922

By the way:

> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/jose-papers/jose.00193/jose.00193/10.21105.jose.00193.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00193/jose.00193/10.21105.jose.00193.pdf) 📄 👈

The PDF itself is needed to be updated, right? or is this an old version a new one is coming?
