### 2024-10-07 #193 @Alberto-GS
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
>
>     * [x]  Double check authors and affiliations (including ORCIDs)
>
>     * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSE paper.
>
>     * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.
>
>     * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSE paper.
>
>     * [x]  Make sure that the license listed for the archive is the same as the software license.

Hello @labarba @kls2177 , checklist double-checked from our side:) What would be the next step?
