### 2024-10-07 #193 @kls2177
Hi @Alberto-GS, there was an error generating the pdf of paper.md. Can you please check this document and confirm that the version that you want generated in the one in the master directory. If it is in a different directory, please specify and I will edit above.
