### 2024-10-07 #193 @PedroVelez
hi @kls2177 , it is the paper.md in the master branch. I have just verified the style and it is the same as in previous version. From the log in the execution : _"/opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych/parser.rb:62:in `_native_parse': (tmp/1de37deeac96e2d979a764c47c3139f3/paper.md): did not find expected key while parsing a block mapping at line 2 column 1 (Psych::SyntaxError)"_ I can not find anything wrong on the paper. md file.

> Hi @Alberto-GS, there was an error generating the pdf of paper.md. Can you please check this document and confirm that the version that you want generated in the one in the master directory. If it is in a different directory, please specify and I will edit above.
