### 2024-10-07 #193 @kls2177
Hi @PedroVelez, I think it has something to do with the indentation. Please refer to this [template](https://openjournals.readthedocs.io/en/jose/submitting.html#example-paper-and-bibliography)
