+++ paper.bib
@@ -1,9 +1,3 @@
-%% This BibTeX bibliography file was created using BibDesk.
-%% https://bibdesk.sourceforge.io/
-%% Created for Alberto Gonzalez at 2023-07-24 16:20:50 +0100 
-%% Saved with string encoding Unicode (UTF-8) 
-
-
 @techreport{ArgoManual2022,
 	title = "Argo user’s manual",
 	type = "Report",
@@ -22,7 +16,8 @@
 	title = {An Introduction to Earth and Environmental Data Science},
 	url = {https://github.com/earth-env-data-science/earth-env-data-science-book},
 	year = {2021},
-	bdsk-url-1 = {https://github.com/earth-env-data-science/earth-env-data-science-book}}
+	bdsk-url-1 = {https://github.com/earth-env-data-science/earth-env-data-science-book}
+}
 
 @misc{jmshea2021,
 	author = {A. M. Smith and K. Thaney and M. Hahnel},
@@ -31,7 +26,8 @@
 	title = {An interactive Quiz generator for Jupyter notebooks and Jupyter Book},
 	url = {https://github.com/jmshea/jupyterquiz},
 	year = {2021},
-	bdsk-url-1 = {https://github.com/jmshea/jupyterquiz}}
+	bdsk-url-1 = {https://github.com/jmshea/jupyterquiz}
+}
 
 @misc{Argo2021,
 	author = {Argo},
@@ -40,7 +36,8 @@
 	title = {Argo float data and metadata from Global Data Assembly Centre (Argo GDAC)},
 	url = {https://doi.org/10.17882/42182},
 	year = {2021},
-	bdsk-url-1 = {https://doi.org/10.17882/42182}}
+	bdsk-url-1 = {https://doi.org/10.17882/42182}
+}
 
 @software{JupyterBook2020,
 	author = {Executable Books Community},
@@ -51,7 +48,8 @@
 	url = {https://doi.org/10.5281/zenodo.4539666},
 	version = {v0.10},
 	year = 2020,
-	bdsk-url-1 = {https://doi.org/10.5281/zenodo.4539666}}
+	bdsk-url-1 = {https://doi.org/10.5281/zenodo.4539666}
+}
 
 @article{roemmich-2019,
[…]
