### 2024-10-07 #193 @kls2177
@PedroVelez, it looks like the line:

`index: 1`

still does not have the right indentation.
