+++ paperNUEVO.md
@@ -1,0 +1,103 @@
+---
+title: 'The Argo Online School: An e-learning tool to get started with Argo'
+tags:
+  - Jupyter Notebooks
+  - Oceanography
+  - Operational oceanography
+  - Observing systems
+  - Robotics
+authors:
+  - name: Alberto González-Santana
+    orcid: 0000-0001-5781-9330
+    affiliation: 1
+  - name: Pedro Vélez-Belchí
+    orcid: 0000-0003-2404-5679
+    affiliation: 1
+affiliations:
+ - name: Centro Oceanográfico de Canarias, Instituto Español de Oceanografía (IEO - CSIC)
+   index: 1
+date: 7 October 2024
+bibliography: paper.bib
+
+---
+
+# Summary
+
+The **Argo Online School (AoS)** is a set of videos, animations and hands-on python driven jupyter notebooks designed to make the data from the more than 4000 profiling floats that constitute the Argo program, accessible. The Argo program samples, in near real time, the upper 2000 meters of the ocean using a fleet of floats that drift with the ocean currents. The AoS is integrated by 28 chapters about the Argo program and its data, organized into a total of 3 lessons and 2 self-assessment sections:
+
+* Lesson 1 is intended to introduce the basics of Argo, its objectives and key elements, such as the structure of the Argo floats and their operation in the open ocean.
+
+* Lesson 2 focuses on the data coming from the Argo network, from its organization to its accessibility. Data quality control is also addressed through its two main modes: Real-Time and Delayed-Mode.
+
+* Lesson 3 is the AoS hands-on component, and it requires basic knowledge of python. It offers a set of instructions to prepare a python environment in case the user wants to run the python Jupyter Notebooks in a local machine. This environment already includes recommended packages. The walkthrough of Lesson 3 shows how to work with the netCDF format, how to access and process Argo data, and how to produce fancy plots for understanding information from Argo data.
+
+The target audience of the AoS are high school, undergraduate or early graduate students, and the programming content included in Lesson 3 provides an ideal opportunity to support students with technical or science curriculum. Lessons 1 and 2 are themselves a closed module and could be used by a learner without using lesson 3, assuming the goal is simply to learn about Argo.  It is recommended not to skip any lesson in the AoS as the content is carefully connected from less to greater complexity, guaranteeing a progressive learning. The user has the opportunity to self-assess the learning process carried out during their course in the AoS through the proposed quizzes.
+
+The AoS has been designed to have the possibility to be expanded to follow the implementation of new features in the Argo program.
+
+# Statement of need
+
+Argo [@roemmich-2019; @riser-2016] is an international program that collects information from inside the ocean using a fleet of floats that drift with the ocean currents. The floats move up and down between the surface and a mid-water level measuring ocean variables and spend almost all their life below the surface. Argo has been collecting more than 100,000 oceanographic profiles per year since 2012 and is composed of approximately 4,000 floats, constituting the major component of the [Global Ocean Observing System](https://www.goosocean.org/).
+
[…]
