### 2024-10-07 c030e95
remover references from paper.md
