### 2024-10-07 #193 @PedroVelez
@kls2177 the error was to include the references in paper.md, since they are in paper.bib .
