### 2024-10-07 0c34422
edited doi references
