+++ paper.bib
@@ -100,7 +100,8 @@
 	title = {Argo Quality Control Manual for CTD and Trajectory Data},
 	url = {https://doi.org/10.13155/33951},
 	year = {2022},
-	bdsk-url-1 = {https://doi.org/10.13155/33951}
+	doi = {10.13155/33951},
+	bdsk-url-1 = {https://dx.doi.org/10.13155/33951}
 }
 
 @article{Argovis,
@@ -126,7 +127,9 @@
 	organization = {IOS Press},
 	pages = {87 - 90},
 	title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
-	year = {2016}
+	year = {2016},
+	bdsk-url-1 = {https://ebooks.iospress.nl/publication/42900},
+	doi = {10.3233/978-1-61499-649-1-87}
 }
 
 @article{Maze2020,
