### 2024-10-11 #193 @kls2177
Hi @PedroVelez, I made a few minor suggested edits to the paper.md file.
