### 2024-10-11 #193 @labarba
Folks, feel free to check items off the [post-review checklist](https://github.com/openjournals/jose-reviews/issues/193#issuecomment-2392387435), as they are double checked by author/editor.

I noticed that the license on the [Zenodo archive](https://zenodo.org/records/13628922) says CC-BY (Attribution), but the repository shows [CC-BY-SA (share-alike)](https://github.com/euroargodev/argoonlineschool/blob/master/license.md). The licenses should match. Please take a look. Also, consider a pure-text (non-markdown) LICENSE file that will get recognized by GitHub.
