### 2024-10-14 #193 @PedroVelez
> Hi @PedroVelez, I made a few minor suggested edits to the paper.md file.

Thanks @kls2177 , and sorry for the questions, where we can find the minor suggested edits to the paper.md file.?
