+++ license.txt
@@ -1,5 +1,5 @@
 ## License for this book
-All the content in the Argo Online School (AOS) is licensed under the [Creative Commons Attribution-ShareAlike 4.0 International](https://creativecommons.org/licenses/by-sa/4.0/) (CC BY-SA 4.0) license.
+All the content in the Argo Online School (AOS) is licensed under the [Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International](https://creativecommons.org/licenses/by-nc-sa/4.0/) (CC BY-NC-SA 4.0) license.
 
 Terms of Use
 Creative Commons Master Terms of Use
