### 2024-10-14 c255ae7
including @kls2177 suggestions in paper.md
