+++ paper.md
@@ -28,27 +28,27 @@ The **Argo Online School (AoS)** is a collection of videos, animations and hands
 
 * Lesson 2 focuses on the data coming from the Argo network, from its organization to its accessibility. Data quality control is also addressed through its two main modes: Real-Time and Delayed-Mode.
 
-* Lesson 3 is the AoS hands-on component, and it requires basic knowledge of python. It provides a set of instructions for preparing a python environment in case the user wants to run the python Jupyter Notebooks on a local machine. This environment already includes recommended packages. The walkthrough of Lesson 3 shows how to work with the netCDF format, how to access and process Argo data, and create visualizations to enhance understanding of the information derived from Argo data.
+* Lesson 3 is the AoS hands-on component, and it requires basic knowledge of python. It provides a set of instructions for preparing a python environment in case the user wants to run the python Jupyter Notebooks on a local machine. This environment already includes recommended packages. The walk-through of Lesson 3 shows how to work with the netCDF format, how to access and process Argo data, and create visualizations to enhance understanding of the information derived from Argo data.
 
-The target audience of the AoS is high school, undergraduate or early graduate students. The programming content in Lesson 3 offers an ideal opportunity to support students pursuing a technical or science curriculum. Lessons 1 and 2 form a closed module and can be used independently by learners who wish to focus solely on Argo. It is recommended not to skip any lesson in the AoS, as the content is carefully structured from simpler to more complex concepts, ensuring progressive learning. Users have the opportunity to self-assess their learning progress through the proposed quizzes in the AoS.
+The target audience of the AoS is high school, undergraduate or early graduate students. The programming content in Lesson 3 offers an ideal opportunity to support students pursuing a technical or science curriculum. Lessons 1 and 2 form a closed module and can be used independently by learners who wish to focus solely on Argo. It is recommended not to skip any lesson in the AoS, as the content is carefully structured from simpler to more complex concepts, providing a progressive learning experience. Users have the opportunity to self-assess their learning progress through the proposed interactive self-assessments in the AoS.
 
-The AoS has been designed to be scalable to follow the implementation of new features in the Argo program.
+The AoS has been designed to be expanded in the future to follow the implementation of new features in the Argo program.
 
 # Statement of need
 
 Argo [@roemmich-2019; @riser-2016] is an international program that collects information from inside the ocean using a fleet of floats that drift with the ocean currents. The floats move up and down between the surface and a mid-water level, measuring ocean variables, and spend almost their entire lifespan below the surface. Since 2012, Argo has collected over 100,000 oceanographic profiles per year and consists of approximately 4,000 floats, making it the major component of the [Global Ocean Observing System](https://www.goosocean.org/).
 
-Although data from Argo [@Argo2021] is publicly and freely available, it represents a vast dataset with thousands of files. The accompanying documentation can be quite complex, as it must address the various models of floats and sensors, as well as the traceability of quality control for all measurements. The Argo community has always recognized the challenges faced by early users in managing the complex and extensive datasets from the network. In this context, the Argo Online School meets the need for effective communication about the Argo program and how its invaluable dataset can be utilized for various applications.
[…]
