+++ paper.md
@@ -63,9 +63,9 @@ Lessons 1 and 2 are designed for students with knowledge comparable to that acqu
 
 These lessons are aimed at users with minimal or no knowledge of the Argo network, and itself are a closed module and could be used by a learner without using Lesson 3, assuming the goal is simply to learn about Argo. Lessons 1 and 2 contain 32 short videos and 14 chapters and require about 5 hours to completed. Lesson 3 is intended for advanced users, as it requires basic programming skills in Python.However, Lesson 3 follows a step by step approach, to facilitate the transition of users coming from Lessons 1 and 2. The basic recommendations and instructions for configuring the hands-on section of the AoS are also provided, whether the users want to work online or if they want to work on their local computer. Specific python libraries and packages are recommended to guarantee the correct functioning of the AoS. Lesson 3 contains 8 short videos and 14 chapters and requires about 10 hours to be completed, assuming basic knowledge of Python.
 
-The AoS has been developed using markdown, python driven Jupyter notebooks [@Kluyver2016jupyter] and jupyter-book [@JupyterBook2020], open source projects that allow editing control in a clear and easy way, and also permits web-based interactive development environments that contain code, visualizations and texts. It is widely used for data science, and inspired by the course _[An Introduction to Earth and Environmental Data Science](https://earth-env-data-science.github.io/intro.html) [@Abernathey2021]. 
+The AoS has been developed using markdown, python driven Jupyter notebooks [@Kluyver2016jupyter] and jupyter-book [@JupyterBook2020], open source projects that allow editing control in a clear and easy way, and also permits web-based interactive development environments that contain code, visualizations and texts. It is widely used for data science, and inspired by the course [An Introduction to Earth and Environmental Data Science](https://earth-env-data-science.github.io/intro.html) [@Abernathey2021].
 
-The AoS is accessible through the Euro-Argo webpage https://www.euro-argo.eu/argo-online-school, but all the content that makes it up is hosted in a public GitHub repository.: [https://github.com/euroargodev/argoonlineschool](https://github.com/euroargodev/argoonlineschool). As long as the Argo program continues to grow, the AoS will too.This first version contains the basic content of Argo; subsequent versions will showcase the newest aspects of the Argo network, such as biochemical measurements, deep observations and more. Anyone interested in helping to further develop the AoS may open an issue on the public GitHub repository to begin organizing the update.
+The AoS is accessible through the Euro-Argo webpage https://www.euro-argo.eu/argo-online-school, but all the content that makes it up is hosted in a public GitHub repository.: [https://github.com/euroargodev/argoonlineschool](https://github.com/euroargodev/argoonlineschool). As long as the Argo program continues to grow, the AoS will too. This first version contains the basic content of Argo; subsequent versions will showcase the newest aspects of the Argo network, such as biochemical measurements, deep observations and more. Anyone interested in helping to further develop the AoS may open an issue on the public GitHub repository to begin organizing the update.
 
 Given the structure of the AoS, it could be used for educational purposes. For high school students, Lessons 1 and 2 could be a project to identify scientific questions that ocean observations, like those from Argo, may help address while developing essential know-how.  For students with a curriculum that includes programming in Python, other projects could involve finding the seasonal change in surface temperature at a specific ocean and comparing it with changes at 1,000 m or 2,000 m, or explaining the trajectory of a given float, which can be quite challenging.  Teachers are welcome to open an issue in the github repository to get assistance in how to develop new projects.
 
