### 2024-10-14 b7beeeb
Update paper.md
Last changes before publishing
