+++ paper.md
@@ -22,13 +22,13 @@ bibliography: paper.bib
 
 # Summary
 
-The **Argo Online School (AoS)** is a collection of videos, animations and hands-on python-driven jupyter notebooks designed to make the data accessible from more than 4,000 profiling floats that constitute the Argo program. The Argo program samples, in near real-time, the upper 2,000 meters of the ocean using a fleet of floats that drift with the ocean currents. The AoS consists of 28 chapters about the Argo program and its data, organized into a total of 3 lessons and 2 self-assessment sections:
+The **Argo Online School (AoS)** is a collection of videos, animations and hands-on Python-driven Jupyter notebooks designed to make the data accessible from more than 4,000 profiling floats that constitute the Argo program. The Argo program samples, in near real-time, the upper 2,000 meters of the ocean using a fleet of floats that drift with the ocean currents. The AoS consists of 28 chapters about the Argo program and its data, organized into a total of 3 lessons and 2 self-assessment sections:
 
 * Lesson 1 is intended to introduce the basics of Argo, its objectives and key elements, such as the structure of the Argo floats and their operation in the open ocean.
 
 * Lesson 2 focuses on the data coming from the Argo network, from its organization to its accessibility. Data quality control is also addressed through its two main modes: Real-Time and Delayed-Mode.
 
-* Lesson 3 is the AoS hands-on component, and it requires basic knowledge of python. It provides a set of instructions for preparing a python environment in case the user wants to run the python Jupyter Notebooks on a local machine. This environment already includes recommended packages. The walk-through of Lesson 3 shows how to work with the netCDF format, how to access and process Argo data, and create visualizations to enhance understanding of the information derived from Argo data.
+* Lesson 3 is the AoS hands-on component, and it requires basic knowledge of Python. It provides a set of instructions for preparing a Python environment in case the user wants to run the Python Jupyter Notebooks on a local machine. This environment already includes recommended packages. The walk-through of Lesson 3 shows how to work with the netCDF format, how to access and process Argo data, and create visualizations to enhance understanding of the information derived from Argo data.
 
 The target audience of the AoS is high school, undergraduate or early graduate students. The programming content in Lesson 3 offers an ideal opportunity to support students pursuing a technical or science curriculum. Lessons 1 and 2 form a closed module and can be used independently by learners who wish to focus solely on Argo. It is recommended not to skip any lesson in the AoS, as the content is carefully structured from simpler to more complex concepts, providing a progressive learning experience. Users have the opportunity to self-assess their learning progress through the proposed interactive self-assessments in the AoS.
 
@@ -46,24 +46,24 @@ The AoS was presented at the [2nd Ocean Observers Workshop](https://bit.ly/3pUCh
 
 # Notes on instructional design
 
-The AoS aims to teach the fundamental concepts needed to understand and use Argo data. It does not seek to cover every aspect of the Argo program, as comprehensive documentation is available from the Argo Data Management Team ([http://www.argodatamgt.org/Documentation](http://www.argodatamgt.org/Documentation)) [@ArgoManual2022, @ArgoQCManual2022] for deeper learning. The AoS is also not library or Application Programming Interface to ease Argo data access, as other initiatives already fulfill that purpose, among others, *Argopy*, a python library for Argo data beginners and experts [@Maze2020] or  *ArgoVis*, a web application for fast delivery, visualization, and analysis of argo Data [@Argovis].
[…]
