### 2024-10-14 #193 @PedroVelez
I haver update the archive version int zenodo to cope with all the changes the new doi is https://doi.org/10.5281/zenodo.13929139 and the new version is 1.1.1 , please @kls2177 could you update both.
