### 2024-10-15 #193 @kls2177
Hi @PedroVelez, one final correction to the references. Please use double brackets for the JupyterBook author:

author = {{Executable Books Community}}

The double brackets won't shorten the author name.
