+++ paper.bib
@@ -40,7 +40,7 @@
 }
 
 @software{JupyterBook2020,
-	author = author = {{Executable Books Community}},
+	author = {{Executable Books Community}},
 	doi = {10.5281/zenodo.4539666},
 	month = feb,
 	publisher = {Zenodo},
