+++ paper.bib
@@ -3,7 +3,7 @@
 	type = "Report",
 	url = "https://doi.org/10.13155/29825",
 	doi = "10.13155/29825",
-	author = {{"Argo data management"}},
+	author = {{Argo data management Team}},
 	year = "2022",
 	editor = "Ifremer",
 	abstract = "This document is the Argo data user’s manual. It contains the description of the formats and files produced by the Argo Data Assembly Centres (DACs)."
