### 2024-10-16 #193 @labarba
@openjournals/dev 👋 We have some Invalid JATS v1.3 error happening here. Eek!
