### 2024-10-17 3a09398
eliminate duplicated reference un paper.bib
