+++ paper.bib
@@ -78,20 +78,6 @@
 	bdsk-url-1 = {http://dx.doi.org/10.1038/nclimate2872}
 }
 
-@article{riser-2016,
-	author = {Riser, Stephen C. and Freeland, Howard J. and Roemmich, Dean and Wijffels, Susan and Troisi, Ariel and Belbeoch, Mathieu and Gilbert, Denis and Xu, Jianping and Pouliquen, Sylvie and Thresher, Ann and Le Traon, Pierre-Yves and Maze, Guillaume and Klein, Birgit and Ravichandran, M. and Grant, Fiona and Poulain, Pierre-Marie and Suga, Toshio and Lim, Byunghwan and Sterl, Andreas and Sutton, Philip and Mork, Kjell-Arne and Velez-Belchi, Pedro Joaquin and Ansorge, Isabelle and King, Brian and Turton, Jon and Baringer, Molly and Jayne, Steven R.},
-	doi = {10.1038/nclimate2872},
-	journal = {Nature Clim. Change},
-	month = {02},
-	number = {2},
-	pages = {145--153},
-	title = {Fifteen years of ocean observations with the global {Argo} array},
-	url = {http://dx.doi.org/10.1038/nclimate2872},
-	volume = {6},
-	year = {2016},
-	bdsk-url-1 = {http://dx.doi.org/10.1038/nclimate2872}
-}
-
 @techreport{ArgoQCManual2022,
 	author = {Wong Annie, Keeley Robert, Carval Thierry},
 	institution = {Argo},
