### 2024-10-17 #193 @xuanxu
It looks like there was a duplicate bib entry, but [has been removed](https://github.com/euroargodev/argoonlineschool/commit/3a093980762cfc8cee2f2acb75939856f609ba68).
It should work now.
