### 2024-10-17 #193 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
