+++ README.md
@@ -1,3 +1,3 @@
 |<img src="https://raw.githubusercontent.com/euroargodev/argoonlineschool/master/images/logoAoS.png" alt="argopy logo" width="100"/> The ``Argo Online School``  is a set of videos and hands-on jupyter notebooks to describe the minimum requiremets to use, and understand, the Argo data. This is the source repository for https://github.com/euroargodev/argoonlineschool/|
 |:---------:|
-|[![status](https://jose.theoj.org/papers/b66eaed8751b3adb6f2f4ad146380818/status.svg)](https://jose.theoj.org/papers/b66eaed8751b3adb6f2f4ad146380818) [![Jupyter Book Badge](https://jupyterbook.org/badge.svg)](<github.com/euroargodev/argoonlineschool/>) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13929139.svg)](https://doi.org/10.5281/zenodo.13929139)| 
+|[![status](https://jose.theoj.org/papers/b66eaed8751b3adb6f2f4ad146380818/status.svg)](https://jose.theoj.org/papers/b66eaed8751b3adb6f2f4ad146380818) [![Jupyter Book Badge](https://jupyterbook.org/badge.svg)](<[github.com/euroargodev/argoonlineschool](https://euroargodev.github.io/argoonlineschool/intro.html)/>) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13929139.svg)](https://doi.org/10.5281/zenodo.13929139)| 
