### 2024-10-20 #193 @labarba
Congrats, @Alberto-GS, your paper s published! 🚀

Huge thanks to our Editor: @kls2177, and Reviewers: @francispoulin, @djmallum — your contributions make this possible 🙏
