### 2024-10-21 #193 @Alberto-GS
> Congrats, @Alberto-GS, your paper s published! 🚀
>
> Huge thanks to our Editor: @kls2177, and Reviewers: @francispoulin, @djmallum — your contributions make this possible 🙏

Thanks a lot!
